% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_celltype_metrics.R
\name{annotate_celltype_metrics}
\alias{annotate_celltype_metrics}
\title{Annotate a SingleCellExperiment With Cell-type Metrics}
\usage{
annotate_celltype_metrics(
  sce,
  cluster_var = "clusters",
  celltype_var = "cluster_celltype",
  unique_id_var = "manifest",
  facet_vars = c("manifest", "group", "sex"),
  input_reduced_dim = "UMAP_Liger",
  metric_vars = c("pc_mito", "pc_ribo", "total_counts", "total_features_by_counts"),
  ...
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{cluster_var}{the colData variable with the cluster numbers}

\item{celltype_var}{the colData variable with the celltype alias}

\item{unique_id_var}{the colData variable with the unique sample name}

\item{facet_vars}{the colData variable(s) for grouped analyses}

\item{input_reduced_dim}{the reducedDim slot used for clustering}

\item{metric_vars}{the numeric colData variable(s) for metric comparisons}

\item{...}{additional parameters}
}
\value{
sce a annotated SingleCellExperiment object
}
\description{
Annotate a SingleCellExperiment With Cell-type Metrics
}
\seealso{
Other annotation functions: 
\code{\link{.preprocess_seurat_object}()},
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce_cells}()},
\code{\link{annotate_sce_genes}()},
\code{\link{annotate_sce}()},
\code{\link{filter_sce}()},
\code{\link{find_cells}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{map_ensembl_gene_id}()},
\code{\link{merge_sce}()},
\code{\link{read_metadata}()},
\code{\link{report_celltype_metrics}()},
\code{\link{report_celltype_model}()},
\code{\link{report_merged_sce}()},
\code{\link{report_qc_sce}()},
\code{\link{run_doubletfinder}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
