% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_sce.R
\name{annotate_sce}
\alias{annotate_sce}
\title{Annotate a SingleCellExperiment With Gene Names and QC Metrics}
\usage{
annotate_sce(
  sce,
  min_library_size = 300,
  max_library_size = "adaptive",
  min_features = 100,
  max_features = "adaptive",
  max_mito = "adaptive",
  type_mito = "both",
  min_ribo = 0,
  max_ribo = 1,
  min_counts = 2,
  min_cells = 2,
  drop_unmapped = TRUE,
  drop_mito = TRUE,
  drop_ribo = FALSE,
  annotate_genes = TRUE,
  annotate_cells = TRUE,
  nmads = 4,
  ensembl_mapping_file = NULL,
  species = getOption("scflow_species", default = "human")
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{min_library_size}{the minimum number of counts per cell}

\item{max_library_size}{the maximum number of counts per cell or "adaptive"}

\item{min_features}{the minimum number of features per cell (i.e. the minimum
number of genes with >0 counts)}

\item{max_features}{the maximum number of features per cell or "adaptive"}

\item{max_mito}{the maximum proportion of counts mapping to
mitochondrial genes (0 - 1) or "adaptive"}

\item{type_mito}{see \code{\link[scater:reexports]{scater::isOutlier()}}. By default, both ends of the
distribution of mitochondrial content will be tested for outliers. If using
single-nucleus RNA-sequencing data, consider setting this to "higher".}

\item{min_ribo}{the minimum proportion of counts mapping to
ribosomal genes (0 - 1)}

\item{max_ribo}{the maximum proportion of counts mapping to
ribosomal genes (0 - 1)}

\item{min_counts}{the minimum number of counts per cell in min_cells}

\item{min_cells}{the minimum number of cells with min_counts}

\item{drop_unmapped}{set \code{TRUE} to remove unmapped ensembl_gene_id}

\item{drop_mito}{set \code{TRUE} to remove mitochondrial genes}

\item{drop_ribo}{set \code{TRUE} to remove ribosomal genes}

\item{annotate_genes}{optionally skip gene annotation with FALSE}

\item{annotate_cells}{optionally skip cell annotation with FALSE}

\item{nmads}{The number of median absolute deviations used to define
outliers for adaptive thresholding.}

\item{ensembl_mapping_file}{a local tsv file with ensembl_gene_id and
additional columns for mapping ensembl_gene_id to gene info.  If
not provided, the biomaRt db is queried (slower).}

\item{species}{The biological species of the sample.}
}
\value{
sce a annotated SingleCellExperiment object
}
\description{
Adds biomaRt annotations (e.g. gene, gene_biotype) and QC metric annotations.
}
\section{Quality control options and thresholds}{


In addition to calculating QC metrics and annotating gene information, this
function adds boolean (TRUE/FALSE) indicators of which cells/genes met the QC
criteria.  This enables QC reports, plots, and various QC-related tables to
be saved before filtering with the  \code{\link[=filter_sce]{filter_sce()}} function.
}

\section{Annotations}{


With the default settings, the SingleCellExperiment object is annotated with:

Cell-level annotations
\itemize{
\item total_counts - sum of counts across all genes
\item total_features_by_counts - total number of unique genes with expression >0
\item qc_metric_min_library_size - did the cell have at least min_library_size
counts
\item qc_metric_min_features - did the cell have counts >0 in at least
min_features number of cells?
\item pc_mito - percentage of counts mapping to mitochondrial genes in this cell
\item qc_metric_pc_mito_ok was pc_mito <= the max_mito cutoff?
\item pc_ribo - percentage of counts mapping to ribosomal genes in this cell
\item qc_metric_pc_ribo_ok was pc_ribo <= the max_ribo cutoff?
\item qc_metric_passed - did the cell pass all of the cell QC tests
}

Gene-level annotations
\itemize{
\item gene - official gene name
\item gene_biotype - protein_coding, lncRNA, pseudogene, etc.
\item qc_metric_ensembl_mapped - was the ensembl_gene_id found in biomaRt
\item qc_metric_is_mito - is the gene mitochondrial
\item qc_metric_is_ribo - is the gene ribosomal
\item qc_metric_n_cells_expressing - number of cells with at least min_counts
\item qc_metric_is_expressive - did at least min_cells have min_counts?
}
}

\seealso{
Other annotation functions: 
\code{\link{.preprocess_seurat_object}()},
\code{\link{annotate_celltype_metrics}()},
\code{\link{annotate_integrated_sce}()},
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce_cells}()},
\code{\link{annotate_sce_genes}()},
\code{\link{filter_sce}()},
\code{\link{find_cells}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{map_ensembl_gene_id}()},
\code{\link{merge_sce}()},
\code{\link{read_metadata}()},
\code{\link{report_celltype_metrics}()},
\code{\link{report_celltype_model}()},
\code{\link{report_merged_sce}()},
\code{\link{report_qc_sce}()},
\code{\link{run_doubletfinder}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
