% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_sce_cells.R
\name{annotate_sce_cells}
\alias{annotate_sce_cells}
\title{Add basic cell-wise annotations for a SingleCellExperiment}
\usage{
annotate_sce_cells(sce, ls_threshold = 300, tf_threshold = 100,
  mt_threshold = 0.1)
}
\arguments{
\item{sce}{a SingleCellExperiment object}
}
\value{
sce a SingleCellExperiment object annotated with cell data
}
\description{
Calculates the following QC metrics:
\itemize{
\item qc_metric_total_counts - was the ensembl_gene_id found in biomaRt
\item qc_metric_total_features_by_counts - is the gene mitochondrial
}
}
\seealso{
Other annotation functions: \code{\link{annotate_sce_genes}},
  \code{\link{generate_sce}},
  \code{\link{map_ensembl_gene_id}},
  \code{\link{retrieve_sample_metadata}}
}
\concept{annotation functions}
