% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_sce_genes.R
\name{annotate_sce_genes}
\alias{annotate_sce_genes}
\title{Add basic gene-wise annotations for a SingleCellExperiment}
\usage{
annotate_sce_genes(sce, ensembl_mapping_file = NULL)
}
\arguments{
\item{sce}{a SingleCellExperiment object with ensembl_gene_id rowData}

\item{ensembl_mapping_file}{a local tsv file with ensembl_gene_id and
additional columns for mapping ensembl_gene_id to gene info.  If
not provided, the biomaRt db is queried (slower).}
}
\value{
sce a SingleCellExperiment object annotated with gene data
}
\description{
Adds biomaRt annotations (e.g. gene, gene_biotype) and calculates the
following QC metrics:
\itemize{
\item qc_metric_ensembl_mapped - was the ensembl_gene_id found in biomaRt
\item qc_metric_is_mito - is the gene mitochondrial
\item qc_metric_is_ribosomal - is the gene ribosomal
}
}
\seealso{
Other annotation functions: \code{\link{annotate_sce_cells}},
  \code{\link{annotate_sce}}, \code{\link{filter_sce}},
  \code{\link{find_singlets}}, \code{\link{generate_sce}},
  \code{\link{map_ensembl_gene_id}},
  \code{\link{merge_sce}}, \code{\link{report_qc_sce}},
  \code{\link{retrieve_sample_metadata}},
  \code{\link{run_doubletfinder}}
}
\concept{annotation functions}
