% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sce.R
\name{cluster_sce}
\alias{cluster_sce}
\title{Cluster SingleCellExperiment with monocle3::cluster_cells}
\usage{
cluster_sce(
  sce,
  cluster_method = "louvain",
  reduction_method = "UMAP_Liger",
  resolution = NULL,
  k = 50,
  louvain_iter = 1,
  verbose = T,
  ...
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{cluster_method}{Available methods include leiden and louvain.}

\item{reduction_method}{Input slot for clustering. Available options are PCA,
UMAP_Liger, tSNE_Liger.}

\item{resolution}{Clustering resolution. If NULL, clustering method will be
set to louvain.}

\item{k}{Integer number of nearest neighbors to use when creating the k
nearest neighbor graph for Louvain/Leiden clustering.
k is related to the resolution of the clustering result, a bigger k will
result in lower resolution and vice versa. Default is 50.}

\item{louvain_iter}{Integer number of iterations used for Louvain clustering.
The clustering result giving the largest modularity score will be used
as the final clustering result. Default is 1. Note that if num_iter
is greater than 1, the random_seed argument will be ignored
for the louvain method.}

\item{verbose}{A logic flag to determine whether or not we should print
the run details.}

\item{...}{see monocle3::cluster_cells for more clustering options.}
}
\value{
sce a SingleCellExperiment object annotated with reducedDims
}
\description{
Cluster SingleCellExperiment with monocle3::cluster_cells
}
\seealso{
Other clustering and dimensionality reduction: 
\code{\link{.get_x_most_specific_genes}()},
\code{\link{.map_celltypes}()},
\code{\link{.map_celltypes_with_ewce}()},
\code{\link{map_custom_celltypes}()},
\code{\link{reduce_dims_sce}()}
}
\concept{clustering and dimensionality reduction}
