% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_with_ewce.R
\name{.append_celltype_plots_sce}
\alias{.append_celltype_plots_sce}
\title{Helper function to plot celltypes in reduced dimensions}
\usage{
.append_celltype_plots_sce(sce, celltype_dim = "cluster_celltype")
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{celltype_dim}{the colData variable with the celltype annotation}
}
\value{
sce a SCE with plots appended to the metadata
}
\description{
Helper function to plot celltypes in reduced dimensions
}
\seealso{
Other helper functions: 
\code{\link{.check_sce_for_liger}()},
\code{\link{.read_rds_files_to_list}()}
}
\concept{helper functions}
\keyword{internal}
