% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{.append_citation_sce}
\alias{.append_citation_sce}
\title{Append a citation to a SingleCellExperiment}
\usage{
.append_citation_sce(sce, key_var = "MENDELEY.TAGS", key, verbose = TRUE)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{key_var}{the key variable (e.g. "MENDELEY.TAGS")}

\item{key}{the value(s) for the key_var (e.g. "doubletfinder")}
}
\value{
sce a SingleCellExperiment object
}
\description{
Append a citation to a SingleCellExperiment
}
\seealso{
Other citation functions: 
\code{\link{.print_citation_html}()},
\code{\link{.print_citation}()}
}
\concept{citation functions}
\keyword{internal}
