% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_webgestaltr.R
\name{.barplot_gsea}
\alias{.barplot_gsea}
\title{barplot for GSEA. x axis normalizedEnrichmentScore, y axis description}
\usage{
.barplot_gsea(dt)
}
\description{
barplot for GSEA. x axis normalizedEnrichmentScore, y axis description
}
\keyword{internal}
