% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_webgestalt.R
\name{.barplot_ora}
\alias{.barplot_ora}
\title{barplot for ORA. x axis enrichmentRatio, y axis description}
\usage{
.barplot_ora(dt)
}
\description{
barplot for ORA. x axis enrichmentRatio, y axis description
}
\keyword{internal}
