% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cells_sce.R
\name{.calculate_retain_parameter}
\alias{.calculate_retain_parameter}
\title{helper fn - calculate the retain parameter for emptydrops based
on the cellranger m/10 of top n cells approach}
\usage{
.calculate_retain_parameter(sce, expect_cells = 3000)
}
\description{
helper fn - calculate the retain parameter for emptydrops based
on the cellranger m/10 of top n cells approach
}
\keyword{internal}
