% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger_preprocess.R
\name{.check_sce_for_liger}
\alias{.check_sce_for_liger}
\title{Helper function to check SCE before running liger}
\usage{
.check_sce_for_liger(sce, ...)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{...}{args}
}
\value{
return 1 if completed
}
\description{
Helper function to check SCE before running liger
}
\seealso{
Other helper functions: 
\code{\link{.append_celltype_plots_sce}()},
\code{\link{.read_rds_files_to_list}()},
\code{\link{.with_env}()}
}
\concept{helper functions}
\keyword{internal}
