% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_enrichr.R
\name{.dotplot_enrichr}
\alias{.dotplot_enrichr}
\title{dotplot for ORA. x axis perturbation, y axis description}
\usage{
.dotplot_enrichr(dt)
}
\description{
dotplot for ORA. x axis perturbation, y axis description
}
\keyword{internal}
