% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_webgestalt.R
\name{.dotplot_gsea}
\alias{.dotplot_gsea}
\title{dotplot for GSEA. x axis normalizedEnrichmentScore, y axis description}
\usage{
.dotplot_gsea(dt)
}
\description{
dotplot for GSEA. x axis normalizedEnrichmentScore, y axis description
}
\keyword{internal}
