% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_webgestaltr.R
\name{.dotplot_ora}
\alias{.dotplot_ora}
\title{dotplot for ORA. x axis enrichmentRatio, y axis description}
\usage{
.dotplot_ora(dt)
}
\description{
dotplot for ORA. x axis enrichmentRatio, y axis description
}
\keyword{internal}
