% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_rontotools.R
\name{.dotplot_pe}
\alias{.dotplot_pe}
\title{dotplot for ORA. x axis perturbation, y axis description}
\usage{
.dotplot_pe(dt)
}
\description{
dotplot for ORA. x axis perturbation, y axis description
}
\keyword{internal}
