% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_doubletfinder.R
\name{.doublet_finder_plot_dim_red}
\alias{.doublet_finder_plot_dim_red}
\title{plot 2d dimensionality reduction with is_doublet colour}
\usage{
.doublet_finder_plot_dim_red(sce, reduced_dim = NULL)
}
\description{
plot 2d dimensionality reduction with is_doublet colour
}
\keyword{internal}
