% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_doubletfinder.R
\name{.doublet_finder_plot_param_sweep}
\alias{.doublet_finder_plot_param_sweep}
\title{pparam sweep plot, pK vs BCmetricx with BCmetric maxima highlighted}
\usage{
.doublet_finder_plot_param_sweep(sce, bcmvn)
}
\description{
pparam sweep plot, pK vs BCmetricx with BCmetric maxima highlighted
}
\keyword{internal}
