% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce.R
\name{.expand_sce_rows}
\alias{.expand_sce_rows}
\title{Helper function to expand a SingleCellExperiment to include specified genes}
\usage{
.expand_sce_rows(sce, union_ensembl_gene_id)
}
\arguments{
\item{sce}{a SingleCellExperiment}

\item{union_ensembl_gene_id}{the full list of ensembl_gene_id to be included
in the new expanded SingleCellExperiment}
}
\value{
sce a annotated SingleCellExperiment object with any
}
\description{
The function identifies the missing ensembl_gene_id's (if any), and
generates new zero'd matrix rows, rbinds them to the matrix, then sorts
the matrix by ensembl_gene_id.  rowData (except ensembl_gene_id) and metadata
are dropped to facilitate subsequent cbind-ing of SingleCellExperiments.
}
\keyword{internal}
