% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp_models.R
\name{.filter_sce_genes_for_de}
\alias{.filter_sce_genes_for_de}
\title{Subset a SingleCellExperiment for genes meeting expressivity criteria}
\usage{
.filter_sce_genes_for_de(sce, min_counts = 1L, min_cells_pc = 0.1)
}
\arguments{
\item{sce}{a SingleCellExperiment object}
}
\value{
sce a SingleCellExperiment object
}
\description{
This function is applied after the subsetting of a SCE (e.g. celltype).
It's useful to subset before differential expression analysis.
}
\seealso{
Other differential gene expression: 
\code{\link{.generate_model_from_vars}()},
\code{\link{.perform_de_with_mast}()},
\code{\link{.preprocess_sce_for_de}()},
\code{\link{perform_de}()},
\code{\link{pseudobulk_sce}()},
\code{\link{report_de}()},
\code{\link{volcano_plot}()}
}
\concept{differential gene expression}
\keyword{internal}
