% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_webgestaltr.R
\name{.format_res_table_ORA}
\alias{.format_res_table_ORA}
\title{Format result table}
\usage{
.format_res_table_ORA(res, enrichment_database)
}
\description{
Format result table
}
\keyword{internal}
