% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_merged_sce.R
\name{.generate_merge_summary_plot}
\alias{.generate_merge_summary_plot}
\title{helper fn to generate a merge summary plot for a SingleCellExperiment}
\usage{
.generate_merge_summary_plot(
  sce,
  plot_var,
  unique_id_var,
  facet_var = NULL,
  plot_points = FALSE
)
}
\arguments{
\item{sce}{a singlecellexperiment object}

\item{plot_var}{the colData variable with values to plot}

\item{unique_id_var}{the colData variable identifying unique samples}

\item{facet_var}{the colData variable to facet/subset by}

\item{plot_points}{if TRUE plot individual values for violin plots}
}
\description{
if plot_var is a factor, a bar plot is returned.
if plot_var is.numeric, a violin plot is returned.
}
\keyword{internal}
