% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp_models.R
\name{.generate_model_from_vars}
\alias{.generate_model_from_vars}
\title{Generate a model formula from dependent, confounding, and random vars}
\usage{
.generate_model_from_vars(
  sce,
  dependent_var,
  confounding_vars,
  random_effects_var = NULL,
  prefix = NULL
)
}
\arguments{
\item{sce}{the data}

\item{dependent_var}{the dependent variable}

\item{confounding_vars}{the confounding variables}

\item{random_effects_var}{the random effect variables (optional)}

\item{prefix}{a variable prefix (e.g. "sce$") (optional)}
}
\value{
model_formula a model formula
}
\description{
Generate a model formula from dependent, confounding, and random vars
}
\seealso{
Other differential gene expression: 
\code{\link{.filter_sce_genes_for_de}()},
\code{\link{.perform_de_with_mast}()},
\code{\link{.preprocess_sce_for_de}()},
\code{\link{perform_de}()},
\code{\link{pseudobulk_sce}()}
}
\concept{differential gene expression}
\keyword{internal}
