% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_merged_sce.R
\name{.generate_pbsce_for_whole_samples}
\alias{.generate_pbsce_for_whole_samples}
\title{helper fn to generate a pseudobulk sce for post-merge qc}
\usage{
.generate_pbsce_for_whole_samples(sce, unique_id_var = "manifest")
}
\arguments{
\item{sce}{a singlecellexperiment object}

\item{unique_id_var}{the colData variable identifying unique samples}
}
\description{
helper fn to generate a pseudobulk sce for post-merge qc
}
\keyword{internal}
