% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_rontotools.R,
%   R/pathway_analysis_webgestaltr.R
\name{.generate_project_name}
\alias{.generate_project_name}
\title{Generating project name}
\usage{
.generate_project_name(gene_file = NULL, enrichment_method = NULL)

.generate_project_name(gene_file = NULL, enrichment_method = NULL)
}
\description{
Generating project name

Generating project name
}
\keyword{internal}
