% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_celltype_metrics.R
\name{.get_d_palette}
\alias{.get_d_palette}
\title{Retrieve a palette of n_colours discrete colours from paletteer}
\usage{
.get_d_palette(pal_name = "ggsci::nrc_npg", n_colours = NULL)
}
\description{
Useful to remove large objects before writing to disk with qs or rds
}
\seealso{
Other helper: 
\code{\link{.append_cell_metric_plots_sce}()},
\code{\link{.append_cell_prop_plots_sce}()},
\code{\link{.clean_ggplot_plot_env}()},
\code{\link{.fancy_scientific}()},
\code{\link{.model_fisher_celltype}()},
\code{\link{.plot_dirichlet_results}()},
\code{\link{.plot_unique_id_var}()},
\code{\link{.prepare_dirichlet_plot_table}()},
\code{\link{.prepare_fisher_counts_table}()},
\code{\link{.prepare_unique_id_var_plot_table}()},
\code{\link{.process_dirichlet_fit}()},
\code{\link{.retrieve_covariates}()},
\code{\link{.tally_cells}()}
}
\concept{helper}
\keyword{internal}
