% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_rontotools.R
\name{.get_overlap_id}
\alias{.get_overlap_id}
\title{Getting the overlapping IDs after enrichment analysis is done}
\usage{
.get_overlap_id(res, res_summary)
}
\value{
Returns the overlapping gene symbols after enrichment analysis
is done
}
\description{
Getting the overlapping IDs after enrichment analysis is done
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{.dotplot_pe}()},
\code{\link{.listdb}()},
\code{\link{find_impacted_pathways}()},
\code{\link{list_databases}()},
\code{\link{pathway_analysis_enrichr}()},
\code{\link{pathway_analysis_rontotools}()},
\code{\link{pathway_analysis_webgestaltr}()},
\code{\link{pathway_to_graph}()},
\code{\link{plot_network}()},
\code{\link{report_impacted_pathway}()}
}
\concept{Impacted pathway analysis}
\keyword{internal}
