% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp_models.R
\name{.get_variance_explained}
\alias{.get_variance_explained}
\title{Get the variance explained by a variable for all genes}
\usage{
.get_variance_explained(
  sce,
  variable = "individual",
  subset_var = NULL,
  subset_class = NULL
)
}
\description{
Get the variance explained by a variable for all genes
}
\keyword{internal}
