% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_rontotools.R
\name{.listdb}
\alias{.listdb}
\title{Getting the list of available database for ROntoTools}
\usage{
.listdb()
}
\description{
Getting the list of available database for ROntoTools
}
\keyword{internal}
