% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_celltype_freqs.R
\name{.model_fisher_celltype}
\alias{.model_fisher_celltype}
\title{Run a fisher exact test on absolute cell numbers for each celltype}
\usage{
.model_fisher_celltype(counts_df, dependent_var, celltype_var, ref_class, ...)
}
\arguments{
\item{counts_df}{absolute cell number counts stratified}

\item{dependent_var}{the name of the colData variable for contrasts}

\item{celltype_var}{the colData variable specifying celltype or subtype}

\item{ref_class}{the class of dependent_var used as reference}
}
\value{
df a data frame of cell numbers with fisher p values adjusted
}
\description{
Run a fisher exact test on absolute cell numbers for each celltype
}
\seealso{
Other helper: 
\code{\link{.append_cell_metric_plots_sce}()},
\code{\link{.append_cell_prop_plots_sce}()},
\code{\link{.clean_ggplot_plot_env}()},
\code{\link{.clean_ggplot_quosures}()},
\code{\link{.clean_quosure_env}()},
\code{\link{.fancy_scientific}()},
\code{\link{.get_d_palette}()},
\code{\link{.grobify_ggplot}()},
\code{\link{.plot_dirichlet_results}()},
\code{\link{.plot_unique_id_var}()},
\code{\link{.prepare_dirichlet_plot_table}()},
\code{\link{.prepare_fisher_counts_table}()},
\code{\link{.prepare_unique_id_var_plot_table}()},
\code{\link{.process_dirichlet_fit}()},
\code{\link{.retrieve_covariates}()},
\code{\link{.tally_cells}()},
\code{\link{.uniftest}()}
}
\concept{helper}
\keyword{internal}
