% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp_models.R
\name{.perform_de_with_mast}
\alias{.perform_de_with_mast}
\title{Perform DE with MAST#'}
\usage{
.perform_de_with_mast(...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}
}
\value{
sce a SingleCellExperiment object
}
\description{
Perform DE with MAST#'
}
\seealso{
Other differential gene expression: \code{\link{.filter_sce_genes_for_de}},
  \code{\link{.generate_model_from_vars}},
  \code{\link{.preprocess_sce_for_de}},
  \code{\link{perform_de}}
}
\concept{differential gene expression}
\keyword{internal}
