% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cells_sce.R
\name{.perform_ks_uniformity_test}
\alias{.perform_ks_uniformity_test}
\title{helper fn - Kolmogorov-Smirnov Uniformity Test}
\usage{
.perform_ks_uniformity_test(x, nrepl = 1000)
}
\description{
helper fn - Kolmogorov-Smirnov Uniformity Test
}
\keyword{internal}
