% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_celltype_freqs.R
\name{.plot_dirichlet_results}
\alias{.plot_dirichlet_results}
\title{Plot relative celltype proportions with Dirichlet model significance}
\usage{
.plot_dirichlet_results(
  df,
  n_groups,
  celltype_var,
  dependent_var,
  facet_plots = FALSE,
  ...
)
}
\description{
Plot relative celltype proportions with Dirichlet model significance
}
\seealso{
Other helper: 
\code{\link{.append_cell_metric_plots_sce}()},
\code{\link{.append_cell_prop_plots_sce}()},
\code{\link{.clean_ggplot_plot_env}()},
\code{\link{.fancy_scientific}()},
\code{\link{.get_d_palette}()},
\code{\link{.model_fisher_celltype}()},
\code{\link{.plot_unique_id_var}()},
\code{\link{.prepare_dirichlet_plot_table}()},
\code{\link{.prepare_fisher_counts_table}()},
\code{\link{.prepare_unique_id_var_plot_table}()},
\code{\link{.process_dirichlet_fit}()},
\code{\link{.retrieve_covariates}()},
\code{\link{.tally_cells}()}
}
\concept{helper}
\keyword{internal}
