% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cells_sce.R
\name{.plot_emptydrops_loglike}
\alias{.plot_emptydrops_loglike}
\title{helper fn - histogram plot to confirm uniform distribution}
\usage{
.plot_emptydrops_loglike(dt)
}
\description{
helper fn - histogram plot to confirm uniform distribution
}
\keyword{internal}
