% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_merged_sce.R
\name{.plot_heatmap_of_pbsce}
\alias{.plot_heatmap_of_pbsce}
\title{helper fn to generate a heatmap and dendrogram for features of a pbsce}
\usage{
.plot_heatmap_of_pbsce(pbsce, binarize = TRUE, trim_name = TRUE)
}
\arguments{
\item{pbsce}{a pseudobulked singlecellexperiment object}

\item{binarize}{if TRUE, counts are TRUE/FALSE if counts > 0}

\item{trim_name}{if TRUE, the pseudobulked name is trimmed to before _}
}
\description{
helper fn to generate a heatmap and dendrogram for features of a pbsce
}
\keyword{internal}
