% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_merged_sce.R
\name{.plot_n_cells_by_unique_id_var}
\alias{.plot_n_cells_by_unique_id_var}
\title{helper fn to generate a cell numbers plot by sample}
\usage{
.plot_n_cells_by_unique_id_var(sce, unique_id_var = "individual")
}
\arguments{
\item{sce}{a singlecellexperiment object}

\item{unique_id_var}{the colData variable identifying unique samples}
}
\description{
helper fn to generate a cell numbers plot by sample
}
\keyword{internal}
