% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_celltype_freqs.R
\name{.plot_unique_id_var}
\alias{.plot_unique_id_var}
\title{Plot relative celltype proportions by sample}
\usage{
.plot_unique_id_var(
  df,
  n_groups,
  celltype_var,
  dependent_var,
  unique_id_var,
  facet_plots = FALSE,
  ...
)
}
\description{
Plot relative celltype proportions by sample
}
\seealso{
Other helper: 
\code{\link{.model_fisher_celltype}()},
\code{\link{.plot_dirichlet_results}()},
\code{\link{.prepare_dirichlet_plot_table}()},
\code{\link{.prepare_fisher_counts_table}()},
\code{\link{.prepare_unique_id_var_plot_table}()},
\code{\link{.process_dirichlet_fit}()},
\code{\link{.retrieve_covariates}()},
\code{\link{.tally_cells}()}
}
\concept{helper}
\keyword{internal}
