% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp_models.R
\name{.preprocess_sce_for_de}
\alias{.preprocess_sce_for_de}
\title{Preprocess the SCE for DE analysis}
\usage{
.preprocess_sce_for_de(...)
}
\arguments{
\item{...}{}
}
\value{
sce a SingleCellExperiment object
}
\description{
Subset genes, set factor reference level, and optionally pseudobulk
}
\seealso{
Other differential gene expression: \code{\link{.filter_sce_genes_for_de}},
  \code{\link{.generate_model_from_vars}},
  \code{\link{.perform_de_with_mast}},
  \code{\link{perform_de}}, \code{\link{pseudobulk_sce}}
}
\concept{differential gene expression}
\keyword{internal}
