% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_doubletfinder.R
\name{.preprocess_seurat_object}
\alias{.preprocess_seurat_object}
\title{temp fn}
\usage{
.preprocess_seurat_object(sce, vars_to_regress_out, pca_dims, var_features)
}
\arguments{
\item{sce}{a SingleCellExperiment object}
}
\value{
seu a seurat object
}
\description{
temp fn
}
\seealso{
Other annotation functions: 
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce_cells}()},
\code{\link{annotate_sce_genes}()},
\code{\link{annotate_sce}()},
\code{\link{filter_sce}()},
\code{\link{find_cells}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{map_ensembl_gene_id}()},
\code{\link{merge_sce}()},
\code{\link{read_metadata}()},
\code{\link{report_celltype_model}()},
\code{\link{report_merged_sce}()},
\code{\link{report_qc_sce}()},
\code{\link{run_doubletfinder}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
\keyword{internal}
