% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{.print_citation}
\alias{.print_citation}
\title{Print a single citation to stdout}
\usage{
.print_citation(citation)
}
\arguments{
\item{citation}{a single row from a bib2df tibble}
}
\description{
Print a single citation to stdout
}
\seealso{
Other citation functions: \code{\link{.append_citation_sce}},
  \code{\link{.print_citation_html}}
}
\concept{citation functions}
\keyword{internal}
