% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{.print_citation_html}
\alias{.print_citation_html}
\title{Print a single citation to stdout as html}
\usage{
.print_citation_html(citation)
}
\arguments{
\item{citation}{a single row from a bib2df tibble}
}
\description{
Print a single citation to stdout as html
}
\seealso{
Other citation functions: 
\code{\link{.append_citation_sce}()},
\code{\link{.print_citation}()}
}
\concept{citation functions}
\keyword{internal}
