% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_celltype_freqs.R
\name{.process_dirichlet_fit}
\alias{.process_dirichlet_fit}
\title{Process the fit model to generate an accessible summary data.frame}
\usage{
.process_dirichlet_fit(fit, dependent_var, celltype_var, ...)
}
\arguments{
\item{fit}{a model generated by \code{DirichletReg::DirichReg}}

\item{dependent_var}{the name of the colData variable for contrasts}

\item{celltype_var}{the colData variable specifying celltype or subtype}
}
\value{
pvals a long format data frame with accessible pvalues
}
\description{
Process the fit model to generate an accessible summary data.frame
}
\seealso{
Other helper: 
\code{\link{.append_cell_metric_plots_sce}()},
\code{\link{.append_cell_prop_plots_sce}()},
\code{\link{.clean_ggplot_plot_env}()},
\code{\link{.clean_ggplot_quosures}()},
\code{\link{.clean_quosure_env}()},
\code{\link{.fancy_scientific}()},
\code{\link{.get_d_palette}()},
\code{\link{.grobify_ggplot}()},
\code{\link{.model_fisher_celltype}()},
\code{\link{.plot_dirichlet_results}()},
\code{\link{.plot_unique_id_var}()},
\code{\link{.prepare_dirichlet_plot_table}()},
\code{\link{.prepare_fisher_counts_table}()},
\code{\link{.prepare_unique_id_var_plot_table}()},
\code{\link{.retrieve_covariates}()},
\code{\link{.tally_cells}()},
\code{\link{.uniftest}()}
}
\concept{helper}
\keyword{internal}
