% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_sce.R
\name{.qc_plot_mito_fraction_histogram}
\alias{.qc_plot_mito_fraction_histogram}
\title{x axis count depth, y axis number of genes}
\usage{
.qc_plot_mito_fraction_histogram(sce)
}
\description{
x axis count depth, y axis number of genes
}
\keyword{internal}
