% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_with_ewce.R
\name{.read_rds_files_to_list}
\alias{.read_rds_files_to_list}
\title{Helper function to read rds files into a list}
\usage{
.read_rds_files_to_list(folder_path)
}
\arguments{
\item{folder_path}{path to folder with rds files}
}
\value{
rds_l list of objects imported with readRDS
}
\description{
Helper function to read rds files into a list
}
\seealso{
Other helper functions: 
\code{\link{.append_celltype_plots_sce}()},
\code{\link{.check_sce_for_liger}()}
}
\concept{helper functions}
\keyword{internal}
