% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_cds.R
\name{.sce_to_cds}
\alias{.sce_to_cds}
\title{Convert a SingleCellExperiment object o a CellDataSet (Monocle)}
\usage{
.sce_to_cds(sce)
}
\arguments{
\item{sce}{a SingleCellExperiment object}
}
\value{
cds a CellDataSet object
}
\description{
Convert a SingleCellExperiment object o a CellDataSet (Monocle)
}
\concept{internal helper}
\keyword{internal}
