% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp_models.R
\name{.volcano_plot}
\alias{.volcano_plot}
\title{volcano plot}
\usage{
.volcano_plot(
  dt = fcHurdle,
  fc_threshold = 2,
  pval_cutoff = 0.001,
  n_label = 10
)
}
\description{
volcano plot
}
\keyword{internal}
