% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_helper_functions.R
\name{.with_env}
\alias{.with_env}
\title{Run function f in environment e}
\usage{
.with_env(f, e = parent.frame())
}
\arguments{
\item{f}{a function}

\item{e}{the environment to run f in}
}
\description{
Run function f in environment e
}
\seealso{
Other helper functions: 
\code{\link{.append_celltype_plots_sce}()},
\code{\link{.check_sce_for_liger}()},
\code{\link{.read_rds_files_to_list}()}
}
\concept{helper functions}
\keyword{internal}
