% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sce.R
\name{filter_sce}
\alias{filter_sce}
\title{Filter SingleCellExperiment according to QC metrics
Quality control metrics previously calculated are used to subset the
SingleCellExperiment for both cells and genes passing QC criteria previously
specified in the \code{annotate_sce_genes} and \code{annotate_sce_cells} functions}
\usage{
filter_sce(sce, filter_genes = FALSE, filter_cells = TRUE,
  keep_mito = FALSE, keep_ribo = TRUE)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{filter_genes}{if set \code{FALSE}, genes will not be filtered}

\item{filter_cells}{if set \code{FALSE}, cells will not be filtered}

\item{keep_mito}{set \code{FALSE} to remove mitochondrial genes}

\item{keep_ribo}{set \code{FALSE} to remove ribosomal genes}
}
\value{
sce a SingleCellExperiment object filtered for QC passed cells and
genes, with QC metrics annotations removed
}
\description{
Filter SingleCellExperiment according to QC metrics
Quality control metrics previously calculated are used to subset the
SingleCellExperiment for both cells and genes passing QC criteria previously
specified in the \code{annotate_sce_genes} and \code{annotate_sce_cells} functions
}
\seealso{
Other annotation functions: \code{\link{annotate_sce_cells}},
  \code{\link{annotate_sce_genes}},
  \code{\link{generate_sce}},
  \code{\link{map_ensembl_gene_id}},
  \code{\link{retrieve_sample_metadata}}
}
\concept{annotation functions}
