% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_marker_genes.R
\name{find_marker_genes}
\alias{find_marker_genes}
\title{Find specific markers for groups of cells}
\usage{
find_marker_genes(
  sce,
  by_vars = c("cluster_celltype", "clusters"),
  fraction_expressing = 0.1,
  top_n = 5,
  max_point_size = 3,
  n_cores = future::availableCores()
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{by_vars}{The colData variable(s) to group cells by}

\item{fraction_expressing}{Top marker genes should be expressed in a minimum
of this fraction of cells (default 0.10)}

\item{top_n}{The top_n genes to use for plotting / subset table generation}

\item{max_point_size}{The point size used for plotting.}

\item{n_cores}{The number of cores to use}
}
\value{
results_l A list of results
}
\description{
Find specific markers for groups of cells
}
\seealso{
Other Further analyses: 
\code{\link{model_celltype_freqs}()}
}
\concept{Further analyses}
