% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sce.R
\name{generate_sce}
\alias{generate_sce}
\title{Generate a SingleCellExperiment object from a matrix and metadata}
\usage{
generate_sce(mat, metadata)
}
\arguments{
\item{mat}{a sparse feature-barcode matrix}

\item{metadata}{a single row dataframe with sample metadata}
}
\value{
sce a SingleCellExperiment object annotated with sample metadata
}
\description{
Generate a SingleCellExperiment with a sparse matrix and a one-row dataframe
containing metadata for the sample.
}
\seealso{
Other annotation functions: \code{\link{annotate_sce_cells}},
  \code{\link{annotate_sce_genes}},
  \code{\link{filter_sce}},
  \code{\link{map_ensembl_gene_id}},
  \code{\link{retrieve_sample_metadata}}
}
\concept{annotation functions}
