% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sce_from_mat_and_meta.R
\name{generate_sce_from_mat_and_meta}
\alias{generate_sce_from_mat_and_meta}
\title{Generate a SingleCellExperiment object from a matrix and metadata}
\usage{
generate_sce_from_mat_and_meta(mat, metadata)
}
\arguments{
\item{mat}{a sparse feature-barcode matrix}

\item{metadata}{a single row dataframe with sample metadata}
}
\value{
sce a SingleCellExperiment object annotated with sample metadata
}
\description{
Generate a SingleCellExperiment with a sparse matrix and a one-row dataframe
containing metadata for the sample.
}
\seealso{
Other annotation functions: \code{\link{map_ensembl_gene_id}},
  \code{\link{retrieve_sample_metadata}}
}
\concept{annotation functions}
