% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_sce.R
\name{integrate_sce}
\alias{integrate_sce}
\title{Integrate datasets within a SingleCellExperiment Object}
\usage{
integrate_sce(sce, method = "Liger", ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object or merged sce objects}

\item{method}{the integration method to use}

\item{...}{Additional arguments}

\item{k}{Inner dimension of factorization (number of factors).
Set to k=30 as default.}
}
\value{
sce SingleCellExperiment object annotated with reducedDims
}
\description{
Produces a reducedDim slot representing common factors across datasets
}
\seealso{
Other Data integration: 
\code{\link{liger_reduce_dims}()}
}
\concept{Data integration}
