% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger_preprocess.R
\name{liger_preprocess}
\alias{liger_preprocess}
\title{Preprocessing steps for Liger dimensionality reduction}
\usage{
liger_preprocess(
  sce,
  unique_id_var = "manifest",
  take_gene_union = F,
  remove.missing = T,
  num_genes = 3000,
  combine = "union",
  keep_unique = F,
  capitalize = F,
  do_plot = F,
  cex_use = 0.3,
  use_cols = T,
  ...
)
}
\arguments{
\item{sce}{merged SingleCellExperiment objects

Make a Liger object:}

\item{remove.missing}{Whether to remove cells from scale.data
with no gene expression (default TRUE)

Remove cells/genes with no expression across any genes/cells:}

\item{combine}{How to combine variable genes across experiments.
Either "union" or "intersect".
(default "union")}

\item{capitalize}{Capitalize gene names to match homologous genes
(ie. across species)
(default FALSE)}

\item{raw.data}{List of expression matrices (gene by cell)
each from an sce object.}

\item{take.gene.union}{Whether to fill out raw.data matrices with union
of genes across all datasets (filling in 0 for missing data)
(requires make.sparse=T) (default FALSE).}

\item{num.genes}{Number of genes to find for each dataset.
Set to 3000 as default.}

\item{keep.unique}{Keep genes that occur (i.e., there is a
corresponding column in raw.data) only in one dataset (default FALSE)}

\item{do.plot}{Display log plot of gene variance vs. gene expression
for each dataset.
Selected genes are plotted in green. (default FALSE)}

\item{cex.use}{Point size for plot.

Scale genes by root-mean-square across cells:}

\item{use.cols}{Treat each column as a cell (default TRUE)}
}
\value{
liger preprocessed object.
}
\description{
Split merged object into multiple sce objects and extract sparse matrices:
}
\seealso{
Other Data integration: 
\code{\link{integrate_sce}()},
\code{\link{liger_reduce_dims}()}
}
\concept{Data integration}
