% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger_reduce_dims.R
\name{liger_reduce_dims}
\alias{liger_reduce_dims}
\title{Dimensionality reduction using Liger factorization}
\usage{
liger_reduce_dims(
  ligerex,
  k = 30,
  lambda = 5,
  method = c("iNMF", "onlineINMF", "UINMF"),
  max_iters = 50,
  h_init = NULL,
  w_init = NULL,
  v_init = NULL,
  rand_seed = 1,
  quantiles = 50,
  ref_dataset = NULL,
  min_cells = 20,
  knn_k = 20,
  center = FALSE,
  ...
)
}
\arguments{
\item{ligerex}{SingleCellExperiment object preprocessed for Liger

Perform iNMF on scaled datasets:}

\item{k}{Inner dimension of factorization (number of factors).
Set to k=30 as default.}

\item{lambda}{Regularization parameter. Larger values penalize
dataset-specific effects more strongly (ie. alignment should increase
as lambda increases). Set to lambda=5.0 as default.}

\item{method}{Method to use for factorization. Options are "iNMF",
"onlineINMF", and "UINMF". (default "iNMF").}

\item{max_iters}{Maximum number of block coordinate descent iterations
to perform (default 100).}

\item{h_init}{Initial values to use for H matrices. (default NULL)}

\item{w_init}{Initial values to use for W matrix (default NULL)}

\item{v_init}{Initial values to use for V matrices (default NULL)}

\item{rand_seed}{Random seed to allow reproducible results (default 1).
Quantile align (normalize) factor loadings:}

\item{quantiles}{Number of quantiles to use for quantile normalization
(default 50).}

\item{ref_dataset}{Name of dataset to use as a "reference"
for normalization. By default, the dataset with the largest number
of cells is used.}

\item{min_cells}{Minimum number of cells to consider a cluster shared
across datasets (default 2)}

\item{knn_k}{Number of nearest neighbors for within-dataset knn graph
(default 20).}

\item{center}{Centers the data when scaling factors (useful for less
sparse modalities like methylation data). (default FALSE)}

\item{...}{Additional arguments.}
}
\value{
liger object with H, H.norm, W, and V slots sets.
}
\description{
Dimensionality reduction using Liger factorization
}
\seealso{
Other Data integration: 
\code{\link{integrate_sce}()},
\code{\link{liger_preprocess}()},
\code{\link{report_integrated_sce}()}
}
\concept{Data integration}
