% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_impacted_pathways.R
\name{list_databases}
\alias{list_databases}
\title{Check available databases and their alias for scFlow}
\usage{
list_databases()
}
\value{
Returns a list containing available database names and their alias for scFlow
}
\description{
Check available databases and their alias for scFlow
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{find_impacted_pathways}()},
\code{\link{pathway_analysis_enrichr}()},
\code{\link{pathway_analysis_rontotools}()},
\code{\link{pathway_analysis_webgestaltr}()},
\code{\link{pathway_to_graph}()},
\code{\link{report_impacted_pathway}()}
}
\concept{Impacted pathway analysis}
