% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_custom_celltypes.R
\name{map_custom_celltypes}
\alias{map_custom_celltypes}
\title{Map Custom Celltype Annotations from a DataFrame}
\usage{
map_custom_celltypes(sce, mappings, cols = NULL, clusters_colname = "clusters")
}
\arguments{
\item{sce}{a SingleCellExperiment}

\item{mappings}{a dataframe with a "cluster" column and additional columns with
celltype annotation data}

\item{cols}{specifies the subset of columns to annotate with.
defaults to NULL or all columns}

\item{clusters_colname}{the name of the colData column with cluster number}
}
\value{
sce a SingleCellExperiment object with custom annotations
}
\description{
Map Custom Celltype Annotations from a DataFrame
}
\seealso{
Other clustering and dimensionality reduction: 
\code{\link{.get_x_most_specific_genes}()},
\code{\link{.map_celltypes_with_ewce}()},
\code{\link{.map_celltypes}()},
\code{\link{cluster_sce}()},
\code{\link{reduce_dims_sce}()}
}
\concept{clustering and dimensionality reduction}
