% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ensembl_gene_id.R
\name{map_ensembl_gene_id}
\alias{map_ensembl_gene_id}
\title{Return mappings for an ensembl gene id}
\usage{
map_ensembl_gene_id(ensembl_ids, mappings = c("ensembl_gene_id",
  "gene_biotype", "external_gene_name"), mappings_filepath = NULL)
}
\arguments{
\item{ensembl_ids}{vector of ensembl_ids, if versioned the version will be
stripped.}

\item{mappings}{the biomaRt attributes to be mapped or mappings file column
names.  This should always include ensembl_gene_id.}

\item{mappings_filepath}{path to the mappings tsv file}
}
\value{
mapped_df a data.frame of the provided ensembl_id's with mappings.
}
\description{
Returns a data frame with the provided ensembl gene ids and requested
mappings from biomaRt or a local file.  If the mappings are absent from the
local file, or the local file is not provided or missing, the mappings are
retrieved by querying biomaRt.
}
\examples{
map_ensembl_gene_id("ENSG00000130707")

}
\concept{annotation functions}
