% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_celltype_freqs.R
\name{model_celltype_freqs}
\alias{model_celltype_freqs}
\title{Model Changes in Celltype Frequencies with Dirichlet Multinomial Regression}
\usage{
model_celltype_freqs(
  sce,
  unique_id_var = "manifest",
  celltype_var = "cluster_celltype",
  dependent_var = "group",
  ref_class = "Control",
  var_order = NULL,
  ...
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{unique_id_var}{the unique identifier variable for each sample}

\item{celltype_var}{the colData variable specifying celltype or subtype}

\item{dependent_var}{the name of the colData variable for contrasts}

\item{ref_class}{the class of dependent_var used as reference}

\item{...}{Additional arguments}
}
\value{
results_l a list of results
}
\description{
Model Changes in Celltype Frequencies with Dirichlet Multinomial Regression
}
\seealso{
Other Further analyses: 
\code{\link{find_marker_genes}()}
}
\concept{Further analyses}
