% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_rontotools.R
\name{pathway_analysis_rontotools}
\alias{pathway_analysis_rontotools}
\title{Functional enrichment analysis using ROntoTools}
\usage{
pathway_analysis_rontotools(
  gene_file = NULL,
  reference_file = NULL,
  project_name = TRUE,
  enrichment_database = "kegg",
  is_output = FALSE,
  output_dir = "."
)
}
\arguments{
\item{gene_file}{A data frame or the path of a .tsv file containing
a list of genes, their fold-change, p-value and adjusted p-value.
Column names should be gene, logFC, pval and padj respectively.}

\item{reference_file}{A .txt file or a data frame containing all the genes
that were used as input for differential expression.
column name should be gene. If not provided the human protein-coding genome
will be used as background genes.}

\item{project_name}{Logical. If TRUE, the gene_file name will be used in
ouput. If FALSE the name will be generate with the timestamp. Default TRUE.}

\item{enrichment_database}{Name of the database for enrichment. User can
specify one or more database names. Default kegg.}

\item{is_output}{If TRUE a folder will be created and results of enrichment
analysis will be saved otherwise a R list will be returned. Default FALSE}

\item{output_dir}{Path for the output directory. Default is current dir.}

\item{plot_output}{If TRUE plots will be saved as .png file otherwise
returned as a R list. Default FALSE.}
}
\value{
enrichment_result a list of data.frames containing enrichment output
and a list of plots of top 10 significant genesets.
}
\description{
Performs impacted pathway analysis with a list of gene, their fold-change
and p-value. The main tool used here is pathway-express (pe).
}
\examples{
set.seed(1234)
enrichment_result <- pathway_analysis_rontotools(
  gene_file = paste(system.file("extdata", package = "scFlowData"), "/",
    "de_result_table.tsv",
    sep = ""
  ),
  project_name = TRUE,
  enrichment_database = "kegg",
  is_output = FALSE
)
}
\seealso{
Other Functional enrichment and impacted pathway analysis: 
\code{\link{pathway_analysis_webgestalt}()}
}
\concept{Functional enrichment and impacted pathway analysis}
