% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_webgestalt.R
\name{pathway_analysis_webgestalt}
\alias{pathway_analysis_webgestalt}
\title{Functional enrichment analysis using ROntoTools}
\usage{
pathway_analysis_webgestalt(
  gene_file = NULL,
  reference_file = NULL,
  enrichment_method = "ORA",
  project_name = TRUE,
  enrichment_database = c("geneontology_Biological_Process_noRedundant",
    "geneontology_Cellular_Component_noRedundant",
    "geneontology_Molecular_Function_noRedundant", "pathway_KEGG", "pathway_Panther",
    "pathway_Reactome", "pathway_Wikipathway"),
  additional_enrichment_databse = FALSE,
  is_output = FALSE,
  output_dir = "."
)
}
\arguments{
\item{gene_file}{A data frame or the path of a .tsv file containing
a list of genes, their fold-change, p-value and adjusted p-value. Column
names should be gene, logFC, pval and padj respectively.}

\item{reference_file}{A data frame containing all the genes that were used
as input for differential expression. Column name should be gene.
If not provided the human protein-coding genome will be used as background
genes.}

\item{enrichment_method}{Method of enrichment analysis.
Either over-representation analysis (ORA) or (Gene set enrichment analysis)
GSEA.}

\item{project_name}{Logical. If TRUE, the gene_file name will be used in
ouput. If FALSE the name will be generate with the timestamp. Default TRUE.}

\item{enrichment_database}{Name of the database for enrichment.
If NULL then multiple databases will be used or user can specify one or more
database names.Default NULL.}

\item{additional_enrichment_databse}{Logical. If TRUE additional databases
will be loaded. If FALSE only default databases will be used. Default TRUE.}

\item{is_output}{If TRUE a folder will be created and results of enrichment
analysis will be saved otherwise a R list will be returned. Default FALSE}

\item{output_dir}{Path for the output directory. Default is current
directory.}
}
\value{
enrichment_result a list of data.frames containing enrichment output
and a list of plots of top 10 significant genesets.
}
\description{
Performs gene ontology and impacted pathway enrichment analysis with a list
of gene names and their fold-change.
}
\examples{
enrichment_result <- pathway_analysis_webgestalt(
  gene_file = paste(system.file("extdata", package = "scflow"), "/",
    "EN-L5-6-I_MASTZLM__DE_groupLow.tsv",
    sep = ""
  ),
  enrichment_method = "ORA",
  project_name = TRUE,
  additional_enrichment_databse = FALSE,
  is_output = TRUE
)
}
\seealso{
Other Functional enrichment and impacted pathway analysis: 
\code{\link{pathway_analysis_rontotools}()}
}
\concept{Functional enrichment and impacted pathway analysis}
