% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_webgestaltr.R
\name{pathway_analysis_webgestaltr}
\alias{pathway_analysis_webgestaltr}
\title{Functional enrichment analysis using WebgestaltR}
\usage{
pathway_analysis_webgestaltr(
  gene_file = NULL,
  reference_file = NULL,
  organism = getOption("scflow_species", default = "human"),
  enrichment_method = "ORA",
  enrichment_database = c("geneontology_Biological_Process",
    "geneontology_Cellular_Component", "geneontology_Molecular_Function", "pathway_KEGG",
    "pathway_Panther", "pathway_Reactome", "pathway_Wikipathway"),
  is_output = FALSE,
  output_dir = "."
)
}
\arguments{
\item{gene_file}{For ORA, A data frame containing a list of significant genes
with column name \code{gene} or a vector of significant genes. For GSEA a
data frame containing a list of all genes in the analysis, their fold-change,
p-value and adjusted p-value. Column names should be gene, logFC, pval and
padj respectively.}

\item{reference_file}{A data frame containing all the genes that were used
as input for differential expression. Column name should be gene.
If not provided the human protein-coding genome will be used as background
genes.}

\item{organism}{default is human. From WebGestaltR supports 12 organisms,
common choices are "hsapiens" or "mmusculus". Users can use the function
\code{\link[WebGestaltR:listOrganism]{WebGestaltR::listOrganism()}} to check available organisms. Users can also
input others to perform the enrichment analysis for other organisms not
supported by WebGestaltR. For other organisms, users need to provide the
functional categories, interesting list and reference list (for ORA method).
Because WebGestaltR does not perform the ID mapping for the other organisms,
the above data should have the same ID type.}

\item{enrichment_method}{Method of enrichment analysis.
Either over-representation analysis (ORA) or (Gene set enrichment analysis)
GSEA.}

\item{enrichment_database}{Name of the database for enrichment. If not
provided then multiple databases will be used or user can specify one or more
database names from \code{\link[WebGestaltR:listGeneSet]{WebGestaltR::listGeneSet()}}}

\item{is_output}{If TRUE a folder will be created and results of enrichment
analysis will be saved otherwise a R list will be returned. Default FALSE.}

\item{output_dir}{Path for the output directory. Default is current
directory.}
}
\value{
enrichment_result a list of data.frames containing enrichment output
and a list of plots of top 10 significant genesets.
}
\description{
Performs gene ontology and impacted pathway enrichment analysis with a list
of gene names and their fold-change.
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{find_impacted_pathways}()},
\code{\link{list_databases}()},
\code{\link{pathway_analysis_enrichr}()},
\code{\link{report_impacted_pathway}()}
}
\concept{Impacted pathway analysis}
