% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_to_graph.R
\name{pathway_to_graph}
\alias{pathway_to_graph}
\title{Convert pathway database into graphNEL format}
\usage{
pathway_to_graph(database_name)
}
\arguments{
\item{database_name}{Name of the database. i.e. "kegg"}
}
\value{
Saves the converted databases as rds in the extdata folder
}
\description{
This function converts a pathway database into graphNEL format using R
package graphite. Currently available databases are kegg, reactome,
nci and panther.
}
\examples{
pathway_to_graph("kegg")
}
\concept{Impacted pathway analysis}
