% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp_models.R
\name{perform_de}
\alias{perform_de}
\title{Perform Differential Gene Expression on a SingleCellExperiment}
\usage{
perform_de(
  sce,
  de_method = "MASTZLM",
  min_counts = 1,
  min_cells_pc = 0.1,
  rescale_numerics = TRUE,
  dependent_var = "group",
  ref_class = "Control",
  confounding_vars = c("individual", "cngeneson", "sex", "age", "PMI", "RIN",
    "seqdate", "pc_mito"),
  random_effects_var = NULL,
  fc_threshold = 1.1,
  pval_cutoff = 0.05,
  ...
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{min_counts}{minimum number of counts}

\item{min_cells_pc}{percentage of cells with min_counts for gene selection}

\item{rescale_numerics}{rescaling numerics may improve model}

\item{dependent_var}{the name of the colData variable for contrasts}

\item{ref_class}{the class of dependent_var used as reference}

\item{confounding_vars}{the independent variables of the model}

\item{random_effects_var}{variable(s) to model as random effects}

\item{fc_threshold}{fold change up/down threshold}

\item{pval_cutoff}{the adjusted pvalue cutoff threshold}

\item{...}{advanced options}
}
\value{
results_l a list of DE table results
}
\description{
Perform Differential Gene Expression on a SingleCellExperiment
}
\seealso{
Other differential gene expression: 
\code{\link{.filter_sce_genes_for_de}()},
\code{\link{.generate_model_from_vars}()},
\code{\link{.perform_de_with_mast}()},
\code{\link{.preprocess_sce_for_de}()},
\code{\link{pseudobulk_sce}()},
\code{\link{report_de}()}
}
\concept{differential gene expression}
