% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expr_by_numeric_var.r
\name{plot_expr_by_numeric_var}
\alias{plot_expr_by_numeric_var}
\title{Plot Gene Expression for Samples against Numerical Variable}
\usage{
plot_expr_by_numeric_var(
  sce,
  numeric_var = "p_tau",
  subset_var = "cluster_celltype",
  subset_group = "Oligo",
  gene = "PLP1",
  alpha = 0.05,
  size = 0.01,
  label_angle = 0
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{numeric_var}{The colData variable for x-axis groups. Default is p_tau}

\item{subset_var}{The colData variable to subset on}

\item{subset_group}{The specific subset_var group to subset}

\item{gene}{The gene of interest}

\item{alpha}{point alpha}

\item{size}{point size}

\item{label_angle}{The angle of x-axis labels (e.g. 0, 45)}
}
\value{
p a ggplot object
}
\description{
Equivalent of plot_violin with a numeric x-axis.
}
\details{
Generates a scatter plot of log2(counts+1) expression against numeric var.
An additional plot of proportion of cells with non-zero counts per sample.
Fits a linear model with confidence intervals for each plot.
}
\seealso{
Other plotting functions: 
\code{\link{plot_reduced_dim}()},
\code{\link{plot_reduced_dim_gene}()},
\code{\link{plot_violin}()}
}
\concept{plotting functions}
