% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot ipa result as network plot}
\usage{
plot_network(enrichment_result, de_table = NULL, show_category = 5, ...)
}
\arguments{
\item{enrichment_result}{A dataframe containing pathway enrichment analysis.}

\item{de_table}{A dataframe containg the gene name and foldchange of the
genes. Ideally the input file for the impacted pathway analysis.}

\item{show_category}{No or names of pathways to see in the netwrok plot.}

\item{...}{Additional arguments}
}
\value{
A network plot object.
}
\description{
Plot ipa result as network plot
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{.dotplot_pe}()},
\code{\link{.get_overlap_id}()},
\code{\link{.listdb}()},
\code{\link{find_impacted_pathways}()},
\code{\link{list_databases}()},
\code{\link{pathway_analysis_enrichr}()},
\code{\link{pathway_analysis_rontotools}()},
\code{\link{pathway_analysis_webgestaltr}()},
\code{\link{pathway_to_graph}()},
\code{\link{report_impacted_pathway}()}
}
\concept{Impacted pathway analysis}
