% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduced_dim.R
\name{plot_reduced_dim}
\alias{plot_reduced_dim}
\title{Plot Cells Reduced Dimensions With Feature Highlighting}
\usage{
plot_reduced_dim(
  sce,
  feature_dim = "Cluster",
  reduced_dim = "UMAP",
  highlight_feature = NA,
  label_clusters = FALSE,
  size = getOption("scflow_reddimplot_pointsize", default = 0.1),
  alpha = getOption("scflow_reddimplot_alpha", default = 0.2)
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{feature_dim}{the colData variable of interest}

\item{reduced_dim}{the reducedDim slot for plotting}

\item{highlight_feature}{highlights a feature}

\item{label_clusters}{prints cluster labels}

\item{size}{point size}

\item{alpha}{point alpha}
}
\value{
p a ggplot object
}
\description{
Generates a 2d plot of cells with feature highlighting
}
\seealso{
Other plotting functions: 
\code{\link{plot_expr_by_numeric_var}()},
\code{\link{plot_reduced_dim_gene}()},
\code{\link{plot_violin}()}
}
\concept{plotting functions}
