% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduced_dim_gene.R
\name{plot_reduced_dim_gene}
\alias{plot_reduced_dim_gene}
\title{Plot Cells Reduced Dimensions With Gene Expression}
\usage{
plot_reduced_dim_gene(
  sce,
  reduced_dim = "UMAP",
  gene = "PLP1",
  size = 0.1,
  alpha = 0.2,
  palette = c("grey80", "#440154FF")
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{reduced_dim}{the reducedDim slot for plotting}

\item{gene}{the gene of interest}

\item{size}{point size}

\item{alpha}{point alpha}

\item{palette}{off and on colours}
}
\value{
p a ggplot object
}
\description{
Plots the cells in reduced dimensionality space with log10 expression values
for a specified gene.
}
\details{
Generates a 2d plot of cells with gene expression
}
\seealso{
Other plotting functions: 
\code{\link{plot_reduced_dim}()},
\code{\link{plot_violin}()}
}
\concept{plotting functions}
