% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_sce.R
\name{pseudobulk_sce}
\alias{pseudobulk_sce}
\title{Generate a Pseudo-bulk SingleCellExperiment from a SingleCellExperiment}
\usage{
pseudobulk_sce(
  sce,
  sample_var = "individual",
  celltype_var = "cluster_celltype",
  assay_name = "counts",
  keep_vars = c("individual", "group", "sex", "age", "PMI", "RIN", "seqdate")
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{sample_var}{the variable name identifying unique samples}

\item{celltype_var}{the variable name identifying cell types}

\item{assay_name}{the assay name for pseudobulking (e.g. "counts")}

\item{keep_vars}{metadata variables to keep from the SingleCellExperiment}
}
\value{
pb_sce a pseudobulk SingleCellExperiment object
}
\description{
Generate a pseudo-bulk of a SingleCellExperiment by summing counts of all
cells from the same \code{sample_var} and \code{celltype_var}.  The SCE is then
re-annotated and library sizes calculated based on cell numbers.
}
\seealso{
Other differential gene expression: 
\code{\link{.filter_sce_genes_for_de}()},
\code{\link{.generate_model_from_vars}()},
\code{\link{.perform_de_with_mast}()},
\code{\link{.preprocess_sce_for_de}()},
\code{\link{perform_de}()},
\code{\link{report_de}()},
\code{\link{volcano_plot}()}
}
\concept{differential gene expression}
