% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_celltype_mappings.R
\name{read_celltype_mappings}
\alias{read_celltype_mappings}
\title{Read Celltype Mappings}
\usage{
read_celltype_mappings(file_path)
}
\arguments{
\item{file_path}{path to read the tsv}
}
\description{
Read Celltype Mappings
}
\seealso{
Other import and export functions: 
\code{\link{read_sce}()},
\code{\link{read_sparse_matrix}()},
\code{\link{write_celltype_mappings}()},
\code{\link{write_sce}()},
\code{\link{write_sparse_matrix}()}
}
\concept{import and export functions}
