% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feature_barcode_matrix.R
\name{read_feature_barcode_matrix}
\alias{read_feature_barcode_matrix}
\title{Read a feature-barcode matrix}
\usage{
read_feature_barcode_matrix(folder_path)
}
\arguments{
\item{folder_path}{path to folder containing feature-barcode matrix files}
}
\value{
mat a sparse matrix with rownames and colnames
}
\description{
Reads a feature-barcode matrix folder into a sparse matrix.  This folder is
output by cellranger and must contain a 'barcodes.tsv.gz', 'features.tsv.gz',
and 'matrix.mtx.gz' file.
}
\seealso{
Other import and export functions: \code{\link{read_sce}},
  \code{\link{write_feature_barcode_matrix}},
  \code{\link{write_sce}}
}
\concept{import and export functions}
