% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read the metadata for a sample from a samplesheet}
\usage{
read_metadata(unique_key, key_colname, samplesheet_path, ...)
}
\arguments{
\item{unique_key}{unique value for this sample in the \code{key_colname} column}

\item{key_colname}{the column name containing the \code{unique_key}}

\item{samplesheet_path}{path to a tsv file containing a samplesheet}

\item{...}{for col_classes (may be extended later - fix)}
}
\value{
metadata the metadata for the specified sample#'
}
\description{
Retrieves a single row from a samplesheet table where the provided
\code{unique_key} parameter is matched in the \code{key_colname} column of the
samplesheet.
}
\seealso{
Other annotation functions: \code{\link{.preprocess_seurat_object}},
  \code{\link{annotate_sce_cells}},
  \code{\link{annotate_sce_genes}},
  \code{\link{annotate_sce}}, \code{\link{filter_sce}},
  \code{\link{find_singlets}}, \code{\link{generate_sce}},
  \code{\link{map_ensembl_gene_id}},
  \code{\link{merge_sce}}, \code{\link{report_qc_sce}},
  \code{\link{run_doubletfinder}}, \code{\link{sce_to_seu}}
}
\concept{annotation functions}
