% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sce.R
\name{read_sce}
\alias{read_sce}
\title{Read a SingleCellExperiment from a folder}
\usage{
read_sce(folder_path, read_metadata = FALSE)
}
\arguments{
\item{folder_path}{path to save the SingleCellExperiment}

\item{read_metadata}{Enable import of metadata if previously saved.}
}
\value{
sce a SingleCellExperiment object
}
\description{
Read a \linkS4class{SingleCellExperiment} from a folder: -
\itemize{
\item Feature-barcode-matrix files: -
\itemize{
\item 'barcodes.tsv.gz'
\item 'features.tsv.gz'
\item 'matrix.mtx.gz'
}
\item Metadata: -
\itemize{
\item 'sce_rowdata.tsv'
\item 'sce_coldata.tsv'
\item 'scecoldata_classes.tsv'
}
\item reducedDim(sce, x) (if present): -
\itemize{
\item 'ReducedDim_x.tsv'
}
}
}
\seealso{
Other import and export functions: 
\code{\link{read_celltype_mappings}()},
\code{\link{read_sparse_matrix}()},
\code{\link{write_celltype_mappings}()},
\code{\link{write_sce}()},
\code{\link{write_sparse_matrix}()}
}
\concept{import and export functions}
