% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_celltype_model.R
\name{report_celltype_model}
\alias{report_celltype_model}
\title{Generate a report for a celltype proportion analysis}
\usage{
report_celltype_model(
  results,
  report_folder_path = getwd(),
  report_file = "celltype_model_report_scflow"
)
}
\arguments{
\item{report_folder_path}{folder path to save the report}

\item{report_file}{filename for report (without an extension)}

\item{sce}{a SingleCellExperiment object}
}
\description{
Generate a report for a celltype proportion analysis
}
\seealso{
Other annotation functions: 
\code{\link{.preprocess_seurat_object}()},
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce_cells}()},
\code{\link{annotate_sce_genes}()},
\code{\link{annotate_sce}()},
\code{\link{filter_sce}()},
\code{\link{find_cells}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{map_ensembl_gene_id}()},
\code{\link{merge_sce}()},
\code{\link{read_metadata}()},
\code{\link{report_merged_sce}()},
\code{\link{report_qc_sce}()},
\code{\link{run_doubletfinder}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
