% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_integrated_sce.R
\name{report_integrated_sce}
\alias{report_integrated_sce}
\title{Generate a report for dataset integration, dims reduction, and clustering}
\usage{
report_integrated_sce(
  sce,
  report_folder_path = getwd(),
  report_file = "integrate_reduceDims_cluster_report_scflow",
  categorical_covariates = list()
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{report_folder_path}{folder path to save the report}

\item{report_file}{filename for report (without an extension)}

\item{categorical_covariates}{list of categorical variables}
}
\value{
sce SingleCellExperiment object annotated with reducedDims
}
\description{
Generate a report for dataset integration, dims reduction, and clustering
}
\seealso{
Other integration, dimension reduction, and clustering: 
\code{\link{annotate_integrated_sce}()}
}
\concept{integration, dimension reduction, and clustering}
