% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_merged_sce.R
\name{report_merged_sce}
\alias{report_merged_sce}
\title{Generate plots and a QC report for a SingleCellExperiment}
\usage{
report_merged_sce(
  sce,
  report_folder_path = getwd(),
  report_file = "merged_report_scflow"
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{report_folder_path}{folder path to save the report}

\item{report_file}{filename for report (without an extension)}
}
\value{
sce a annotated SingleCellExperiment object
}
\description{
Generate plots and a QC report for a SingleCellExperiment
}
\seealso{
Other annotation functions: 
\code{\link{.preprocess_seurat_object}()},
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce_cells}()},
\code{\link{annotate_sce_genes}()},
\code{\link{annotate_sce}()},
\code{\link{filter_sce}()},
\code{\link{find_cells}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{map_ensembl_gene_id}()},
\code{\link{merge_sce}()},
\code{\link{read_metadata}()},
\code{\link{report_celltype_model}()},
\code{\link{report_qc_sce}()},
\code{\link{run_doubletfinder}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
