% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_qc_sce.R
\name{report_qc_sce}
\alias{report_qc_sce}
\title{Generate plots and a QC report for a SingleCellExperiment}
\usage{
report_qc_sce(sce)
}
\arguments{
\item{sce}{a SingleCellExperiment object}
}
\value{
sce a annotated SingleCellExperiment object
}
\description{
Generate plots and a QC report for a SingleCellExperiment
}
\seealso{
Other annotation functions: \code{\link{annotate_sce_cells}},
  \code{\link{annotate_sce_genes}},
  \code{\link{annotate_sce}}, \code{\link{filter_sce}},
  \code{\link{find_singlets}}, \code{\link{generate_sce}},
  \code{\link{map_ensembl_gene_id}},
  \code{\link{merge_sce}},
  \code{\link{retrieve_sample_metadata}},
  \code{\link{run_doubletfinder}}
}
\concept{annotation functions}
