% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_sample_metadata.R
\name{retrieve_sample_metadata}
\alias{retrieve_sample_metadata}
\title{Retrieve the metadata for a sample from a samplesheet}
\usage{
retrieve_sample_metadata(unique_id, id_colname, samplesheet_path, ...)
}
\arguments{
\item{unique_id}{unique value for this sample in the \code{id_colname} column}

\item{id_colname}{the column name containing the \code{unique_id}}

\item{samplesheet_path}{path to a tsv file containing a samplesheet}

\item{...}{for colClasses (may be extended later - fix)}
}
\value{
metadata the metadata for the specified sample
}
\description{
Retrieves a single row from a samplesheet table where the provided
\code{unique_id} parameter is matched in the \code{id_colname} column of the
samplesheet.
}
\seealso{
Other annotation functions: \code{\link{annotate_sce_cells}},
  \code{\link{annotate_sce_genes}},
  \code{\link{annotate_sce}}, \code{\link{filter_sce}},
  \code{\link{find_singlets}}, \code{\link{generate_sce}},
  \code{\link{map_ensembl_gene_id}},
  \code{\link{merge_sce}}, \code{\link{report_qc_sce}},
  \code{\link{run_doubletfinder}}
}
\concept{annotation functions}
