% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_doubletfinder.R
\name{run_doubletfinder}
\alias{run_doubletfinder}
\title{Find singlets in a SingleCellExperiment with DoubletFinder by McGinnis CS
et al}
\usage{
run_doubletfinder(
  sce,
  pK = NULL,
  pca_dims = 10,
  var_features = 2000,
  vars_to_regress_out = "nCount_RNA",
  doublet_rate = 0,
  dpk = 8,
  num_cores = max(1, future::availableCores() - 2)
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{pK}{a pK value to use in place of a parameter sweep}

\item{pca_dims}{the number of principal components to use}

\item{var_features}{the top n variable features to use}

\item{vars_to_regress_out}{the variables to regress out}

\item{doublet_rate}{either a fixed doublet rate (e.g. 0.075) or 0 to}

\item{dpk}{doublets per thousand cells increment if doublet_rate is 0.}

\item{num.cores}{the number of CPU cores to use}
}
\value{
sce a SingleCellExperiment object annotated for singlets
}
\description{
Runs the doubletfinder algorithm on the SingleCellExperiment. Returns a
SingleCellExperiment annotated with is_singlet and tSNE coordinates.
}
\details{
Remember to cite: -
DoubletFinder: Doublet Detection in Single-Cell RNA Sequencing Data Using
Artificial Nearest Neighbors.
McGinnis CS, Murrow LM, Gartner ZJ.
Cell Syst. 2019 Apr 24;8(4):329-337.e4.
doi: \url{https://doi.org/10.1016/j.cels.2019.03.003}.
}
\seealso{
Other annotation functions: 
\code{\link{.preprocess_seurat_object}()},
\code{\link{annotate_celltype_metrics}()},
\code{\link{annotate_integrated_sce}()},
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce}()},
\code{\link{annotate_sce_cells}()},
\code{\link{annotate_sce_genes}()},
\code{\link{filter_sce}()},
\code{\link{find_cells}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{map_ensembl_gene_id}()},
\code{\link{merge_sce}()},
\code{\link{read_metadata}()},
\code{\link{report_celltype_metrics}()},
\code{\link{report_celltype_model}()},
\code{\link{report_merged_sce}()},
\code{\link{report_qc_sce}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
\keyword{internal}
