% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_seurat.R
\name{sce_to_seu}
\alias{sce_to_seu}
\title{temp fn}
\usage{
sce_to_seu(sce)
}
\arguments{
\item{sce}{a SingleCellExperiment object}
}
\value{
seu a seurat object
}
\description{
temp fn
}
\seealso{
Other annotation functions: \code{\link{.preprocess_seurat_object}},
  \code{\link{annotate_sce_cells}},
  \code{\link{annotate_sce_genes}},
  \code{\link{annotate_sce}}, \code{\link{filter_sce}},
  \code{\link{find_singlets}}, \code{\link{generate_sce}},
  \code{\link{map_ensembl_gene_id}},
  \code{\link{merge_sce}}, \code{\link{report_qc_sce}},
  \code{\link{retrieve_metadata}},
  \code{\link{run_doubletfinder}}
}
\concept{annotation functions}
\keyword{internal}
