% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp_models.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Plot volcano plot for differential expression analysis}
\usage{
volcano_plot(dt, fc_threshold = 1.05, pval_cutoff = 0.05, n_label = 10)
}
\arguments{
\item{dt}{Differential expression result table from perform_de() function.}

\item{fc_threshold}{Fold change threshold for the volcano plot. This will be adjusted and plotted as the log2 fold change.  Default is 1.05.}

\item{pval_cutoff}{The adjusted p-value cut-off for the volcano plot. Default is 0.05.}

\item{n_label}{The number of top up & down differentially expressed genes to be labeled. Default is 10.}
}
\value{
No return. The plot is printed out.
}
\description{
Plot volcano plot for differential expression analysis
}
\seealso{
Other differential gene expression: 
\code{\link{.filter_sce_genes_for_de}()},
\code{\link{.generate_model_from_vars}()},
\code{\link{.perform_de_with_mast}()},
\code{\link{.preprocess_sce_for_de}()},
\code{\link{perform_de}()},
\code{\link{pseudobulk_sce}()},
\code{\link{report_de}()}
}
\concept{differential gene expression}
