% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_celltype_mappings.R
\name{write_celltype_mappings}
\alias{write_celltype_mappings}
\title{Write Celltype Mappings}
\usage{
write_celltype_mappings(
  sce,
  folder_path,
  write_metadata = FALSE,
  overwrite_files = TRUE
)
}
\arguments{
\item{sce}{a singlecellexperiment}

\item{folder_path}{path to save the tsv}

\item{overwrite_files}{set to \code{TRUE} to overwrite files if they already
exist.}
}
\description{
Write Celltype Mappings
}
\seealso{
Other import and export functions: 
\code{\link{read_celltype_mappings}()},
\code{\link{read_sce}()},
\code{\link{read_sparse_matrix}()},
\code{\link{write_sce}()},
\code{\link{write_sparse_matrix}()}
}
\concept{import and export functions}
