% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_feature_barcode_matrix.R
\name{write_feature_barcode_matrix}
\alias{write_feature_barcode_matrix}
\title{Write a feature-barcode matrix}
\usage{
write_feature_barcode_matrix(mat, folder_path, overwrite_files = TRUE)
}
\arguments{
\item{mat}{a matrix}

\item{folder_path}{path to save the feature barcode matrix files}
}
\description{
Write a feature-barcode matrix into a folder with .gz compressed files.  This
folder contains 'barcodes.tsv.gz', 'features.tsv.gz',and 'matrix.mtx.gz'.
}
\seealso{
Other import and export functions: \code{\link{read_feature_barcode_matrix}},
  \code{\link{write_sce}}
}
\concept{import and export functions}
