% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sce.R
\name{write_sce}
\alias{write_sce}
\title{Write a SingleCellExperiment to a folder}
\usage{
write_sce(mat, folder_path)
}
\arguments{
\item{folder_path}{path to save the SingleCellExperiment}

\item{sce}{a SingleCellExperiment Object}
}
\description{
Write a SingleCellExperiment into a folder.  Files saved include: -
\itemize{
\item Feature-barcode-matrix files: -
\itemize{
\item 'barcodes.tsv.gz'
\item 'features.tsv.gz'
\item 'matrix.mtx.gz'
}
\item Metadata: -
\itemize{
\item sce_rowdata.tsv
\item sce_coldata.tsv
}
\item reducedDim(sce, x) (if present): -
\itemize{
\item ReducedDim_x.tsv
}
}
}
\seealso{
Other import and export functions: \code{\link{read_feature_barcode_matrix}},
  \code{\link{write_feature_barcode_matrix}}
}
\concept{import and export functions}
