% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sparse_matrix.R
\name{write_sparse_matrix}
\alias{write_sparse_matrix}
\title{Write a feature-barcode matrix}
\usage{
write_sparse_matrix(mat, folder_path, overwrite_files = TRUE)
}
\arguments{
\item{mat}{a matrix}

\item{folder_path}{path to save the feature barcode matrix files}

\item{overwrite_files}{set to \code{TRUE} to overwrite files if they already
exist.}
}
\description{
Write a feature-barcode matrix into a folder with .gz compressed files.  This
folder contains 'barcodes.tsv.gz', 'features.tsv.gz',and 'matrix.mtx.gz'.
}
\seealso{
Other import and export functions: 
\code{\link{read_celltype_mappings}()},
\code{\link{read_sce}()},
\code{\link{read_sparse_matrix}()},
\code{\link{write_celltype_mappings}()},
\code{\link{write_sce}()}
}
\concept{import and export functions}
