% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_integrated_sce.R
\name{annotate_integrated_sce}
\alias{annotate_integrated_sce}
\title{Annotate integrated, reduced dimension,
and clustered SingleCellExperiment object}
\usage{
annotate_integrated_sce(sce, categorical_covariates = list())
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{categorical_covariates}{list of categorical variables}
}
\value{
sce annotated SingleCellExperiment object
}
\description{
Annotate integrated, reduced dimension,
and clustered SingleCellExperiment object
}
\seealso{
Other integration, dimension reduction, and clustering: 
\code{\link{report_integrated_sce}()}
}
\concept{integration, dimension reduction, and clustering}
