% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_integrated_sce.R
\name{annotate_integrated_sce}
\alias{annotate_integrated_sce}
\title{Annotate integrated, reduced dimension,
and clustered SingleCellExperiment object}
\usage{
annotate_integrated_sce(
  sce,
  categorical_covariates = list(),
  input_reduced_dim = "tSNE"
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{categorical_covariates}{list of categorical variables}

\item{input_reduced_dim}{which reduced dim to use for integration and
clustering. Either "tSNE" or "UMAP" (default "tSNE").}
}
\value{
sce annotated SingleCellExperiment object
}
\description{
Annotate integrated, reduced dimension,
and clustered SingleCellExperiment object
}
\seealso{
Other annotation functions: 
\code{\link{.preprocess_seurat_object}()},
\code{\link{annotate_celltype_metrics}()},
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce}()},
\code{\link{annotate_sce_cells}()},
\code{\link{annotate_sce_genes}()},
\code{\link{filter_sce}()},
\code{\link{find_cells}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{map_ensembl_gene_id}()},
\code{\link{merge_sce}()},
\code{\link{read_metadata}()},
\code{\link{report_celltype_metrics}()},
\code{\link{report_celltype_model}()},
\code{\link{report_merged_sce}()},
\code{\link{report_qc_sce}()},
\code{\link{run_doubletfinder}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
