% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_sce_cells.R
\name{annotate_sce_cells}
\alias{annotate_sce_cells}
\title{Add basic cell-wise annotations for a SingleCellExperiment}
\usage{
annotate_sce_cells(sce, min_library_size = 300, min_features = 100,
  max_mito = 0.1, max_ribo = 1)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{min_library_size}{the minimum number of counts per cell}

\item{min_features}{the minimum number of features per cell (i.e. the minimum
number of genes with >0 counts)}

\item{max_mito}{the maximum proportion of counts mapping to
mitochondrial genes (0 - 1)}

\item{max_ribo}{the maximum proportion of counts mapping to
ribosomal genes (0 - 1)}
}
\value{
sce a SingleCellExperiment object annotated with cell QC metrics
}
\description{
Calculates the following QC metrics:
\itemize{
\item qc_metric_total_counts - was the ensembl_gene_id found in biomaRt
\item qc_metric_total_features_by_counts - is the gene mitochondrial
}
}
\seealso{
Other annotation functions: \code{\link{annotate_sce_genes}},
  \code{\link{generate_sce}},
  \code{\link{map_ensembl_gene_id}},
  \code{\link{retrieve_sample_metadata}}
}
\concept{annotation functions}
