% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_sce_cells.R
\name{annotate_sce_cells}
\alias{annotate_sce_cells}
\title{Add basic cell-wise annotations for a SingleCellExperiment}
\usage{
annotate_sce_cells(sce, ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{...}{See \code{\link{annotate_sce}}

min_library_size the minimum number of counts per cell
min_features the minimum number of features per cell (i.e. the minimum
number of genes with >0 counts)
max_mito the maximum proportion of counts mapping to
mitochondrial genes (0 - 1)
min_ribo
max_ribo the maximum proportion of counts mapping to
ribosomal genes (0 - 1)
min_counts the minimum number of counts per cell in min_cells
min_cells the minimum number of cells with min_counts}
}
\value{
sce a SingleCellExperiment object annotated with cell QC metrics
}
\description{
Calculates the following QC metrics:
\itemize{
\item qc_metric_total_counts - was the ensembl_gene_id found in biomaRt
\item qc_metric_total_features_by_counts - is the gene mitochondrial
}
}
\seealso{
Other annotation functions: 
\code{\link{.preprocess_seurat_object}()},
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce_genes}()},
\code{\link{annotate_sce}()},
\code{\link{filter_sce}()},
\code{\link{find_cells}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{map_ensembl_gene_id}()},
\code{\link{merge_sce}()},
\code{\link{read_metadata}()},
\code{\link{report_impacted_pathway}()},
\code{\link{report_merged_sce}()},
\code{\link{report_qc_sce}()},
\code{\link{run_doubletfinder}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
