% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sce.R
\name{cluster_sce}
\alias{cluster_sce}
\title{Cluster SingleCellExperiment with monocle3::cluster_cells}
\usage{
cluster_sce(sce, ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{...}{see uwot::umap for umap options}

\item{reduction_methods}{one or more of "PCA", "tSNE", "UMAP", "UMAP3D"}

\item{pca_dims}{the number of pca dimensions used}
}
\value{
sce a SingleCellExperiment object annotated with reducedDims
}
\description{
Cluster SingleCellExperiment with monocle3::cluster_cells
}
\seealso{
Other clustering and dimensionality reduction: 
\code{\link{.get_x_most_specific_genes}()},
\code{\link{.map_celltypes_with_ewce}()},
\code{\link{.map_celltypes}()},
\code{\link{map_celltypes_sce}()},
\code{\link{reduce_dims_sce}()}
}
\concept{clustering and dimensionality reduction}
