% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_helper_functions.R
\name{.clean_ggplot_plot_env}
\alias{.clean_ggplot_plot_env}
\title{Remove objects from a ggplot plot_env}
\usage{
.clean_ggplot_plot_env(p, drop_classes = c("SingleCellExperiment", "ggplot"))
}
\description{
Useful to remove large objects before writing to disk with qs or rds
This function was deprecated by p$plot_env <- rlang::new_environment()
}
\seealso{
Other helper: 
\code{\link{.append_cell_metric_plots_sce}()},
\code{\link{.append_cell_prop_plots_sce}()},
\code{\link{.clean_ggplot_quosures}()},
\code{\link{.clean_quosure_env}()},
\code{\link{.fancy_scientific}()},
\code{\link{.get_d_palette}()},
\code{\link{.grobify_ggplot}()},
\code{\link{.model_fisher_celltype}()},
\code{\link{.plot_dirichlet_results}()},
\code{\link{.plot_unique_id_var}()},
\code{\link{.prepare_dirichlet_plot_table}()},
\code{\link{.prepare_fisher_counts_table}()},
\code{\link{.prepare_unique_id_var_plot_table}()},
\code{\link{.process_dirichlet_fit}()},
\code{\link{.retrieve_covariates}()},
\code{\link{.tally_cells}()},
\code{\link{.uniftest}()}
}
\concept{helper}
\keyword{internal}
