% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_rontotools.R
\name{.dotplot_pe}
\alias{.dotplot_pe}
\title{dotplot for ORA. x axis perturbation, y axis description}
\usage{
.dotplot_pe(dt)
}
\description{
dotplot for ORA. x axis perturbation, y axis description
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{.get_overlap_id}()},
\code{\link{.listdb}()},
\code{\link{find_impacted_pathways}()},
\code{\link{list_databases}()},
\code{\link{pathway_analysis_enrichr}()},
\code{\link{pathway_analysis_rontotools}()},
\code{\link{pathway_analysis_webgestaltr}()},
\code{\link{pathway_to_graph}()},
\code{\link{plot_network}()},
\code{\link{report_impacted_pathway}()}
}
\concept{Impacted pathway analysis}
\keyword{internal}
