% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp_models.R
\name{.generate_model_from_vars}
\alias{.generate_model_from_vars}
\title{Generate a model formula from dependent, confounding, and random vars}
\usage{
.generate_model_from_vars(dependent_var, confounding_vars,
  re_vars = NULL)
}
\arguments{
\item{dependent_var}{the dependent variable}

\item{confounding_vars}{the confounding variables}

\item{re_vars}{the random effect variables (optional)}
}
\value{
model_formula a model formula
}
\description{
Generate a model formula from dependent, confounding, and random vars
}
\seealso{
Other differential gene expression: \code{\link{.filter_sce_genes_for_de}},
  \code{\link{.perform_de_with_mast}},
  \code{\link{.preprocess_sce_for_de}},
  \code{\link{perform_de}}
}
\concept{differential gene expression}
\keyword{internal}
