% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_webgestalt.R
\name{.generate_project_name}
\alias{.generate_project_name}
\title{Generating project name}
\usage{
.generate_project_name(
  project_name = NULL,
  gene_file = NULL,
  enrichment_method = NULL
)
}
\description{
Generating project name
}
\keyword{internal}
