% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_rontotools.R
\name{.get_overlap_id}
\alias{.get_overlap_id}
\title{Getting the overlapping IDs after enrichment analysis is done}
\usage{
.get_overlap_id(res, res_summary)
}
\description{
Getting the overlapping IDs after enrichment analysis is done
}
\keyword{internal}
