% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_rontotools.R
\name{.listdb}
\alias{.listdb}
\title{Getting the list of available database for ROntoTools}
\usage{
.listdb()
}
\value{
Returns a data frame of available databases for ROntoTools
}
\description{
Getting the list of available database for ROntoTools
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{.dotplot_pe}()},
\code{\link{.get_overlap_id}()},
\code{\link{find_impacted_pathways}()},
\code{\link{list_databases}()},
\code{\link{pathway_analysis_enrichr}()},
\code{\link{pathway_analysis_rontotools}()},
\code{\link{pathway_analysis_webgestaltr}()},
\code{\link{pathway_to_graph}()},
\code{\link{plot_network}()},
\code{\link{report_impacted_pathway}()}
}
\concept{Impacted pathway analysis}
\keyword{internal}
