% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_sce.R
\name{.qc_plot_count_depth_distribution}
\alias{.qc_plot_count_depth_distribution}
\title{x axis barcode rank, y axis total counts}
\usage{
.qc_plot_count_depth_distribution(sce)
}
\description{
x axis barcode rank, y axis total counts
}
\keyword{internal}
