% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sce.R
\name{filter_sce}
\alias{filter_sce}
\title{Filter SingleCellExperiment according to QC metrics
Quality control metrics previously calculated are used to subset the
SingleCellExperiment for both cells and genes passing QC criteria previously
specified in the \code{annotate_sce_genes} and \code{annotate_sce_cells} functions}
\usage{
filter_sce(sce, filter_genes = TRUE, filter_cells = TRUE)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{filter_genes}{if set \code{FALSE}, genes will not be filtered}

\item{filter_cells}{if set \code{FALSE}, cells will not be filtered}
}
\value{
sce a SingleCellExperiment object filtered for QC passed cells and
genes, with QC metrics annotations removed
}
\description{
Filter SingleCellExperiment according to QC metrics
Quality control metrics previously calculated are used to subset the
SingleCellExperiment for both cells and genes passing QC criteria previously
specified in the \code{annotate_sce_genes} and \code{annotate_sce_cells} functions
}
\seealso{
Other annotation functions: 
\code{\link{.preprocess_seurat_object}()},
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce_cells}()},
\code{\link{annotate_sce_genes}()},
\code{\link{annotate_sce}()},
\code{\link{find_cells}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{map_ensembl_gene_id}()},
\code{\link{merge_sce}()},
\code{\link{read_metadata}()},
\code{\link{report_impacted_pathway}()},
\code{\link{report_merged_sce}()},
\code{\link{report_qc_sce}()},
\code{\link{run_doubletfinder}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
