% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cells_sce.R
\name{find_cells}
\alias{find_cells}
\title{Annotate a SingleCellExperiment with EmptyDrops predictions}
\usage{
find_cells(
  sce,
  lower = 100,
  retain = "auto",
  alpha_cutoff = 0.001,
  niters = 10000,
  expect_cells = 3000,
  ...
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{lower}{see \code{\link[DropletUtils:emptyDrops]{DropletUtils::emptyDrops()}}}

\item{retain}{see \code{\link[DropletUtils:emptyDrops]{DropletUtils::emptyDrops()}}.  Use "auto" to calculate
the parameter from the top \code{expect_cells} cells (cellranger method).}

\item{alpha_cutoff}{The alpha cutoff used for Monte Carlo signficance.}

\item{niters}{the number of Monte Carlo iterations}

\item{expect_cells}{the number of cells expected for auto retain}

\item{...}{additional arguments for uniformity testing}
}
\value{
sce a SingleCellExperiment object annotated with emptyDrops metrics
}
\description{
Calculates the following QC metric:
\itemize{
\item is_empty_drop - is the barcode an empty drop
}
}
\seealso{
Other annotation functions: 
\code{\link{.preprocess_seurat_object}()},
\code{\link{annotate_celltype_metrics}()},
\code{\link{annotate_integrated_sce}()},
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce}()},
\code{\link{annotate_sce_cells}()},
\code{\link{annotate_sce_genes}()},
\code{\link{filter_sce}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{map_ensembl_gene_id}()},
\code{\link{merge_sce}()},
\code{\link{read_metadata}()},
\code{\link{report_celltype_metrics}()},
\code{\link{report_celltype_model}()},
\code{\link{report_merged_sce}()},
\code{\link{report_qc_sce}()},
\code{\link{run_doubletfinder}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
