% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_impacted_pathways.R
\name{find_impacted_pathways}
\alias{find_impacted_pathways}
\title{Functional enrichment analysis}
\usage{
find_impacted_pathways(
  gene_file = NULL,
  reference_file = NULL,
  enrichment_tool = c("WebGestaltR", "ROntoTools", "enrichR"),
  enrichment_database = c("KEGG", "Panther", "Reactome"),
  is_output = FALSE,
  output_dir = ".",
  ...
)
}
\arguments{
\item{gene_file}{A data frame or the path of a .tsv file containing
a list of genes, their fold-change, p-value and adjusted p-value. Column
names should be gene, logFC, pval and padj respectively.}

\item{reference_file}{A data frame containing all the genes that were used
as input for differential expression. Column name should be gene.
If not provided the human protein-coding genome will be used as background
genes.}

\item{enrichment_tool}{Enrichment tool to use. WebGestaltR, ROntoTools and
enrichR is implemented. Use one or more of the tools.}

\item{enrichment_database}{Name of the database for enrichment. User can
specify one or more database names. Check \code{\link[=list_databases]{list_databases()}} for
available database alias.}

\item{is_output}{If TRUE a folder will be created and results of enrichment
analysis will be saved otherwise a R list will be returned. Default FALSE.}

\item{output_dir}{Path for the output directory. Default is current
directory.}

\item{...}{Additional arguments}
}
\value{
enrichment_result a list of list containing enrichment outputs from
different enrichment tools.
}
\description{
Performs gene ontology and impacted pathway enrichment analysis with a list
of gene names and their fold-change.
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{.dotplot_pe}()},
\code{\link{.get_overlap_id}()},
\code{\link{.listdb}()},
\code{\link{list_databases}()},
\code{\link{pathway_analysis_enrichr}()},
\code{\link{pathway_analysis_rontotools}()},
\code{\link{pathway_analysis_webgestaltr}()},
\code{\link{pathway_to_graph}()},
\code{\link{plot_network}()},
\code{\link{report_impacted_pathway}()}
}
\concept{Impacted pathway analysis}
