% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_impacted_pathways.R
\name{find_impacted_pathways}
\alias{find_impacted_pathways}
\title{Functional enrichment analysis}
\usage{
find_impacted_pathways(
  gene_file = NULL,
  reference_file = NULL,
  enrichment_tool = c("WebgestaltR", "ROntoTools"),
  enrichment_database = c("pathway_KEGG", "pathway_Panther", "pathway_Reactome"),
  is_output = FALSE,
  output_dir = ".",
  ...
)
}
\arguments{
\item{gene_file}{A data frame or the path of a .tsv file containing
a list of genes, their fold-change, p-value and adjusted p-value. Column
names should be gene, logFC, pval and padj respectively.}

\item{reference_file}{A data frame containing all the genes that were used
as input for differential expression. Column name should be gene.
If not provided the human protein-coding genome will be used as background
genes.}

\item{enrichment_tool}{Enrichment tool to use. WebGestaltR, ROntoTools and
EnrichR is implemented.}

\item{enrichment_database}{Name of the database for enrichment.
If NULL then multiple databases will be used or user can specify one or more
database names.Default NULL.}

\item{is_output}{If TRUE a folder will be created and results of enrichment
analysis will be saved otherwise a R list will be returned. Default FALSE}

\item{output_dir}{Path for the output directory. Default is current
directory.}
}
\value{
enrichment_result a list of list containing enrichment output from
different enrichment tools and a list of plots.
}
\description{
Performs gene ontology and impacted pathway enrichment analysis with a list
of gene names and their fold-change.
}
\examples{
enrichment_result <- find_impacted_pathways(
  gene_file = paste(system.file("extdata", package = "scFlowData"), "/",
    "de_result_table.tsv",
    sep = ""
  ),
  enrichment_method = "ORA",
  additional_enrichment_databse = FALSE
)
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{pathway_analysis_enrichr}()},
\code{\link{pathway_analysis_rontotools}()},
\code{\link{pathway_analysis_webgestaltr}()},
\code{\link{pathway_to_graph}()}
}
\concept{Impacted pathway analysis}
