% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_singlets.R
\name{find_singlets}
\alias{find_singlets}
\title{Find singlets in a SingleCellExperiment}
\usage{
find_singlets(sce, singlet_find_method)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{singlet_find_method}{the method to use for identifying singlets}
}
\value{
sce a SingleCellExperiment object annotated for singlets
}
\description{
Runs a doublet/multiplet identification algorithm on the SingleCellExperiment
and returns a SingleCellExperiment annotated with is_singlet and, depending
on the algorithm, additional annotations / dimensionality reductions.
}
\seealso{
Other annotation functions: \code{\link{annotate_sce_cells}},
  \code{\link{annotate_sce_genes}},
  \code{\link{annotate_sce}}, \code{\link{filter_sce}},
  \code{\link{generate_sce}},
  \code{\link{map_ensembl_gene_id}},
  \code{\link{retrieve_sample_metadata}},
  \code{\link{run_doubletfinder}}
}
\concept{annotation functions}
