% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_sce.R
\name{integrate_sce}
\alias{integrate_sce}
\title{Integrate datasets within a SingleCellExperiment Object}
\usage{
integrate_sce(sce, method = "Liger", ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object or merged sce objects}

\item{method}{the integration method to use. Set to liger as default.
See \code{liger_preprocess} and \code{liger_reduce_dims} functions for parameter options.}

\item{...}{Additional arguments. Such as k, inner dimension of factorization
(number of factors).}
}
\value{
sce SingleCellExperiment object annotated with reducedDims
}
\description{
Produces a reducedDim slot representing common factors across datasets
}
\seealso{
Other Data integration: 
\code{\link{liger_preprocess}()},
\code{\link{liger_reduce_dims}()},
\code{\link{report_integrated_sce}()}
}
\concept{Data integration}
