% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger_preprocess.R
\name{liger_preprocess}
\alias{liger_preprocess}
\title{Preprocessing steps for Liger dimensionality reduction}
\usage{
liger_preprocess(
  sce,
  k,
  unique_id_var = "manifest",
  take_gene_union = F,
  remove.missing = T,
  num_genes = 2000,
  combine = "union",
  capitalize = F,
  use_cols = T,
  num_cores = 1,
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment object or merged objects}

\item{k}{Inner dimension of factorization (number of factors).}

\item{unique_id_var}{the colData variable identifying unique samples.
Default is "manifest".

Make a Liger object:}

\item{take_gene_union}{Whether to fill out raw.data matrices with union
of genes across all datasets (filling in 0 for missing data)
(requires make.sparse=T) (default FALSE).}

\item{remove.missing}{Whether to remove cells not expressing any
measured genes, and genes not
expressed in any cells (if take.gene.union = T, removes only genes
not expressed in any dataset) (default TRUE).

Select informative genes:}

\item{num_genes}{Number of genes to find for each dataset.
Set to 3000 as default.}

\item{combine}{How to combine variable genes across experiments.
Either "union" or "intersect".
(default "union")}

\item{capitalize}{Capitalize gene names to match homologous genes
(ie. across species)
(default FALSE)
Scale genes by root-mean-square across cells:

Remove cells/genes with no expression across any genes/cells:}

\item{use_cols}{Treat each column as a cell (default TRUE)}

\item{num_cores}{Number of cores used on user's machine to run function.
Default is 1.}

\item{...}{Additional arguments.}
}
\value{
liger preprocessed object.
}
\description{
Split merged object into multiple sce objects and extract sparse matrices:
}
\seealso{
Other Data integration: 
\code{\link{integrate_sce}()},
\code{\link{liger_reduce_dims}()},
\code{\link{report_integrated_sce}()}
}
\concept{Data integration}
