% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger_preprocess.R
\name{liger_preprocess}
\alias{liger_preprocess}
\title{Preprocessing steps for Liger dimensionality reduction}
\usage{
liger_preprocess(
  sce,
  k,
  unique_id_var = "manifest",
  take_gene_union = F,
  remove.missing = T,
  num_genes = 3000,
  combine = "union",
  capitalize = F,
  use_cols = T,
  num_cores = future::availableCores(),
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment object or merged objects}

\item{k}{Inner dimension of factorization (number of factors).

Make a Liger object:}

\item{remove.missing}{Whether to remove cells not expressing any
measured genes, and genes not
expressed in any cells (if take.gene.union = T, removes only genes
not expressed in any dataset) (default TRUE).

Select informative genes:}

\item{combine}{How to combine variable genes across experiments.
Either "union" or "intersect".
(default "union")}

\item{capitalize}{Capitalize gene names to match homologous genes
(ie. across species)
(default FALSE)
Scale genes by root-mean-square across cells:

Remove cells/genes with no expression across any genes/cells:}

\item{take.gene.union}{Whether to fill out raw.data matrices with union
of genes across all datasets (filling in 0 for missing data)
(requires make.sparse=T) (default FALSE).}

\item{num.genes}{Number of genes to find for each dataset.
Set to 3000 as default.}

\item{use.cols}{Treat each column as a cell (default TRUE)}
}
\value{
liger preprocessed object.
}
\description{
Split merged object into multiple sce objects and extract sparse matrices:
}
