% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_with_ewce.R
\name{map_celltypes_sce}
\alias{map_celltypes_sce}
\title{Map cluster celltypes with EWCE}
\usage{
map_celltypes_sce(
  sce,
  ctd_folder,
  cells_to_sample = 10000,
  clusters_colname = "clusters",
  species = "human"
)
}
\arguments{
\item{sce}{a SingleCellExperiment}

\item{ctd_folder}{path to a folder with ctd RDS files}

\item{cells_to_sample}{total cells to consider for gene enrichment}

\item{clusters_colname}{the name of the colData column with cluster number}

\item{species}{specify human or mouse}
}
\value{
sce a SingleCellExperiment object annotated with celltypes/metadata
}
\description{
Currently works with Allan and Zeisel only
}
\seealso{
Other clustering and dimensionality reduction: 
\code{\link{.get_x_most_specific_genes}()},
\code{\link{.map_celltypes_with_ewce}()},
\code{\link{.map_celltypes}()},
\code{\link{cluster_sce}()},
\code{\link{map_custom_celltypes}()},
\code{\link{reduce_dims_sce}()}
}
\author{
Nathan Skene / Combiz Khozoie
}
\concept{clustering and dimensionality reduction}
