% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ensembl_gene_id.R
\name{map_ensembl_gene_id}
\alias{map_ensembl_gene_id}
\title{Return mappings for Ensembl Gene IDs}
\usage{
map_ensembl_gene_id(
  ensembl_ids,
  mappings = c("ensembl_gene_id", "gene_biotype", "external_gene_name",
    "percentage_gene_gc_content"),
  species = getOption("scflow_species", default = "human"),
  ensembl_mapping_file = NULL
)
}
\arguments{
\item{ensembl_ids}{vector of ensembl_ids, if versioned the version will be
stripped.}

\item{mappings}{the biomaRt attributes to be mapped or mappings file column
names.  ensembl_gene_id will be added if not included here.
\code{c("ensembl_gene_id", "external_gene_name")}}

\item{species}{ensembl_ids are mapped to \code{human} or \code{mouse}}

\item{ensembl_mapping_file}{path to the mappings tsv file}
}
\value{
mapped_df a data.frame of the provided ensembl_id's with mappings.
}
\description{
Returns a data frame with the provided ensembl gene ids and requested
mappings from biomaRt or a local file.  If the mappings are absent from the
local file, or the local file is not provided or missing, the mappings are
retrieved by querying biomaRt.
}
\seealso{
Other annotation functions: 
\code{\link{.preprocess_seurat_object}()},
\code{\link{annotate_celltype_metrics}()},
\code{\link{annotate_integrated_sce}()},
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce_cells}()},
\code{\link{annotate_sce_genes}()},
\code{\link{annotate_sce}()},
\code{\link{filter_sce}()},
\code{\link{find_cells}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{merge_sce}()},
\code{\link{read_metadata}()},
\code{\link{report_celltype_metrics}()},
\code{\link{report_celltype_model}()},
\code{\link{report_merged_sce}()},
\code{\link{report_qc_sce}()},
\code{\link{run_doubletfinder}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
