% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce.R
\name{merge_sce}
\alias{merge_sce}
\title{Merge Multiple SingleCellExperiment Objects}
\usage{
merge_sce(sce_l, ensembl_mapping_file = NULL)
}
\arguments{
\item{sce_l}{a list of SingleCellExperiment objects, or, folder_paths to
SingleCellExperiment objects to be read in with \code{\link{read_sce}}}

\item{ensembl_mapping_file}{path to the mappings tsv file}
}
\value{
sce a annotated SingleCellExperiment object
}
\description{
Merge Multiple SingleCellExperiment Objects
}
\seealso{
Other annotation functions: 
\code{\link{.preprocess_seurat_object}()},
\code{\link{annotate_merged_sce}()},
\code{\link{annotate_sce_cells}()},
\code{\link{annotate_sce_empty}()},
\code{\link{annotate_sce_genes}()},
\code{\link{annotate_sce}()},
\code{\link{filter_sce}()},
\code{\link{find_singlets}()},
\code{\link{generate_sce}()},
\code{\link{map_ensembl_gene_id}()},
\code{\link{read_metadata}()},
\code{\link{report_merged_sce}()},
\code{\link{report_qc_sce}()},
\code{\link{run_doubletfinder}()},
\code{\link{sce_to_seu}()}
}
\concept{annotation functions}
