% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_enrichr.R
\name{pathway_analysis_enrichr}
\alias{pathway_analysis_enrichr}
\title{Functional enrichment analysis using enrichR}
\usage{
pathway_analysis_enrichr(
  gene_file = NULL,
  reference_file = NULL,
  enrichment_database = c("GO_Molecular_Function_2018", "GO_Cellular_Component_2018",
    "GO_Biological_Process_2018", "MGI_Mammalian_Phenotype_2017",
    "MGI_Mammalian_Phenotype_Level_3", "MGI_Mammalian_Phenotype_Level_4", "ChEA_2016",
    "ENCODE_and_ChEA_Consensus_TFs_from_ChIP-X", "KEGG_2019_Human",
    "WikiPathways_2019_Human", "Reactome_2016", "Allen_Brain_Atlas_down",
    "Allen_Brain_Atlas_up"),
  is_output = FALSE,
  output_dir = "."
)
}
\arguments{
\item{gene_file}{A data frame or the path of a .tsv file containing
a list of genes, their fold-change, p-value and adjusted p-value.
Column names should be gene, logFC, pval and padj respectively.}

\item{enrichment_database}{Name of the database for enrichment. User can
specify one or more database names from \code{\link[enrichR:listEnrichDbs]{enrichR::listEnrichDbs()}}.}

\item{is_output}{If TRUE a folder will be created and results of enrichment
analysis will be saved otherwise a R list will be returned. Default FALSE}

\item{output_dir}{Path for the output directory. Default is current dir.}

\item{plot_output}{If TRUE plots will be saved as .png file otherwise
returned as a R list. Default FALSE.}
}
\value{
enrichment_result a list of data.frames containing enrichment output
and a list of plots of top 10 significant genesets.
}
\description{
Performs impacted pathway analysis with a list of genes.
}
\examples{
set.seed(1234)
enrichment_result <- pathway_analysis_enrichr(
  gene_file = paste(system.file("extdata", package = "scFlowData"), "/",
    "de_result_table.tsv",
    sep = ""
  ),
  enrichment_database = "GO_Molecular_Function_2018",
  is_output = FALSE
)
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{find_impacted_pathways}()},
\code{\link{pathway_analysis_rontotools}()},
\code{\link{pathway_analysis_webgestaltr}()},
\code{\link{pathway_to_graph}()}
}
\concept{Impacted pathway analysis}
