% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_enrichr.R
\name{pathway_analysis_enrichr}
\alias{pathway_analysis_enrichr}
\title{Functional enrichment analysis using enrichR}
\usage{
pathway_analysis_enrichr(
  gene_file = NULL,
  enrichment_database = c("GO_Molecular_Function_2021", "GO_Cellular_Component_2021",
    "GO_Biological_Process_2021", "KEGG_2021_Human", "WikiPathways_2021_Human",
    "Reactome_2022"),
  is_output = FALSE,
  output_dir = getwd()
)
}
\arguments{
\item{gene_file}{A data frame containing a list of significant genes with
column name gene or a vector of significant genes.}

\item{enrichment_database}{Name of the database for enrichment. User can
specify one or more database names from \code{\link[enrichR:listEnrichrDbs]{enrichR::listEnrichrDbs()}}.}

\item{is_output}{If TRUE a folder will be created and results of enrichment
analysis will be saved otherwise a R list will be returned. Default FALSE}

\item{output_dir}{Path for the output directory. Default is current dir.}
}
\value{
enrichment_result a list of data.frames containing enrichment output
and a list of plots of top 10 significant genesets.
}
\description{
Performs impacted pathway analysis with a list of genes.
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{find_impacted_pathways}()},
\code{\link{list_databases}()},
\code{\link{pathway_analysis_webgestaltr}()},
\code{\link{report_impacted_pathway}()}
}
\concept{Impacted pathway analysis}
