% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_to_graph.R
\name{pathway_to_graph}
\alias{pathway_to_graph}
\title{Convert pathway database into graphNEL format}
\usage{
pathway_to_graph(database_name)
}
\arguments{
\item{database_name}{Name of the database. i.e. "kegg"}
}
\value{
Saves the converted databases as rds in the extdata folder
}
\description{
This function converts a pathway database into graphNEL format using R
package graphite. Currently available databases are kegg, reactome,
nci and panther.
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{.dotplot_pe}()},
\code{\link{.get_overlap_id}()},
\code{\link{.listdb}()},
\code{\link{find_impacted_pathways}()},
\code{\link{list_databases}()},
\code{\link{pathway_analysis_enrichr}()},
\code{\link{pathway_analysis_rontotools}()},
\code{\link{pathway_analysis_webgestaltr}()},
\code{\link{plot_network}()},
\code{\link{report_impacted_pathway}()}
}
\concept{Impacted pathway analysis}
