% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_dimension_plots.R
\name{plot_reduced_dim}
\alias{plot_reduced_dim}
\title{Plot Cells Reduced Dimensions With Feature Highlighting}
\usage{
plot_reduced_dim(sce, feature_dim = "Cluster", reduced_dim = "UMAP",
  highlight_feature = NA, label_clusters = FALSE, size = 0.1,
  alpha = 0.2)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{feature_dim}{the colData variable of interest}

\item{reduced_dim}{the reducedDim slot for plotting}

\item{highlight_feature}{highlights a feature}

\item{label_clusters}{prints cluster labels}
}
\value{
p a ggplot object
}
\description{
Generates a 2d plot of cells with feature highlighting
}
\concept{plotting functions}
