% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dims_sce.R
\name{reduce_dims_sce}
\alias{reduce_dims_sce}
\title{Calculate dimensionality reductions for a SingleCellExperiment}
\usage{
reduce_dims_sce(sce, reduction_methods = c("PCA", "tSNE", "UMAP",
  "UMAP3D"), vars_to_regress_out = NULL, pca_dims = 20, ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{reduction_methods}{one or more of "PCA", "tSNE", "UMAP", "UMAP3D"}

\item{pca_dims}{the number of pca dimensions used}

\item{...}{see uwot::umap for umap options}
}
\value{
sce a SingleCellExperiment object annotated with reducedDims
}
\description{
PCA, tSNE, UMAP, UMAP3D
}
\seealso{
Other clustering and dimensionality reduction: \code{\link{cluster_sce}}
}
\concept{clustering and dimensionality reduction}
