% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_de.R
\name{report_de}
\alias{report_de}
\title{Generate a report for differential expression analysis}
\usage{
report_de(res, report_folder_path = getwd(), report_file = "de_report_scflow")
}
\arguments{
\item{res}{Differential expression result table from perform_de() function.}

\item{report_folder_path}{folder path to save the report.}

\item{report_file}{filename for report (without an extension).}
}
\description{
Generate a report for differential expression analysis
}
\seealso{
Other differential gene expression: 
\code{\link{.filter_sce_genes_for_de}()},
\code{\link{.generate_model_from_vars}()},
\code{\link{.perform_de_with_mast}()},
\code{\link{.preprocess_sce_for_de}()},
\code{\link{perform_de}()},
\code{\link{pseudobulk_sce}()}
}
\concept{differential gene expression}
