% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_impacted_pathway.R
\name{report_impacted_pathway}
\alias{report_impacted_pathway}
\title{Generate a report for impacted pathway analysis}
\usage{
report_impacted_pathway(
  res,
  report_folder_path = getwd(),
  report_file = "ipa_report_scflow"
)
}
\arguments{
\item{res}{Pathway enrichment result from find_impacted_pathway() function.}

\item{report_folder_path}{folder path to save the report.}

\item{report_file}{filename for report (without an extension).}
}
\description{
Generate a report for impacted pathway analysis
}
\seealso{
Other Impacted pathway analysis: 
\code{\link{find_impacted_pathways}()},
\code{\link{list_databases}()},
\code{\link{pathway_analysis_enrichr}()},
\code{\link{pathway_analysis_rontotools}()},
\code{\link{pathway_analysis_webgestaltr}()},
\code{\link{pathway_to_graph}()}
}
\concept{Impacted pathway analysis}
