% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_doubletfinder.R
\name{run_doubletfinder}
\alias{run_doubletfinder}
\title{Find singlets in a SingleCellExperiment with DoubletFinder by McGinnis CS
et al}
\usage{
run_doubletfinder(sce, ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{...}{'pK' passed here will override parameter sweep,
'doublet_rate' passed here will override the default of 7.5\%
a doublet_rate of 0 is a special case that automates the estimated dr}
}
\value{
sce a SingleCellExperiment object annotated for singlets
}
\description{
Runs the doubletfinder algorithm on the SingleCellExperiment. Returns a
SingleCellExperiment annotated with is_singlet and tSNE coordinates.
}
\details{
Remember to cite: -
DoubletFinder: Doublet Detection in Single-Cell RNA Sequencing Data Using
Artificial Nearest Neighbors.
McGinnis CS, Murrow LM, Gartner ZJ.
Cell Syst. 2019 Apr 24;8(4):329-337.e4.
doi: \url{https://doi.org/10.1016/j.cels.2019.03.003}.
}
\seealso{
Other annotation functions: \code{\link{.preprocess_seurat_object}},
  \code{\link{annotate_merged_sce}},
  \code{\link{annotate_sce_cells}},
  \code{\link{annotate_sce_genes}},
  \code{\link{annotate_sce}}, \code{\link{filter_sce}},
  \code{\link{find_singlets}}, \code{\link{generate_sce}},
  \code{\link{map_ensembl_gene_id}},
  \code{\link{merge_sce}}, \code{\link{read_metadata}},
  \code{\link{report_qc_sce}}, \code{\link{sce_to_seu}}
}
\concept{annotation functions}
\keyword{internal}
