% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sce.R
\name{write_sce}
\alias{write_sce}
\title{Write a SingleCellExperiment to a folder}
\usage{
write_sce(sce, folder_path, overwrite = TRUE)
}
\arguments{
\item{sce}{a SingleCellExperiment Object}

\item{folder_path}{path to save the SingleCellExperiment}

\item{overwrite}{delete the existing directory and overwrite if TRUE}
}
\description{
Write a \linkS4class{SingleCellExperiment} into a folder: -
\itemize{
\item Feature-barcode-matrix files: -
\itemize{
\item 'barcodes.tsv.gz'
\item 'features.tsv.gz'
\item 'matrix.mtx.gz'
}
\item Metadata: -
\itemize{
\item sce_rowdata.tsv
\item sce_coldata.tsv
\item scecoldata_classes.tsv
}
\item reducedDim(sce, x) (if present): -
\itemize{
\item ReducedDim_x.tsv
}
}
}
\seealso{
Other import and export functions: \code{\link{read_sce}},
  \code{\link{read_sparse_matrix}},
  \code{\link{write_sparse_matrix}}
}
\concept{import and export functions}
